@@ -11,16 +11,16 @@
     "lint": "eslint ts/ --ext .ts,.tsx",
     "lint-fix": "eslint ts/ --ext .ts,.tsx --fix",
     "compile": "./node_modules/.bin/tsc",
-    "prepare": "npm run clean; node tests.config.js; npm run compile; npm run webpack",
+    "prepare": "npm run clean; node tests.config.js; npm run compile",
     "webpack": "./node_modules/.bin/webpack",
     "test": "npm run test:files",
     "test:all": "npm run test:files; npm run test:output; npm run test:analytics",
     "test:files": "npx jest --testPathIgnorePatterns js/output --testPathIgnorePatterns js/analyse",
     "test:output": "npx jest --testPathPattern js/output",
     "test:analytics": "npx rimraf analysis; npx jest --testPathPattern js/analyse; npm run test:allrules",
-    "test:allrules": "echo \"import AnalyticsTest from './js/analytics/analytics_test'; AnalyticsTest.outputAllRules()\" | node -r esm",
-    "analyse:tries": "echo \"import AnalyticsTrie from './js/analytics/analytics_trie'; AnalyticsTrie.output()\" | node -r esm",
-    "clean": "npx rimraf dist js ts/output ts/json ts/analyse coverage"
+    "test:allrules": "echo \"import('./js/module_loader.js').then((m) => m.loadPromise('at')).then(m => m.default.outputAllRules())\" | node ",
+    "analyse:tries": "echo \"import('./js/module_loader.js').then((m) => m.loadPromise('atrie')).then(m => m.default.output())\" | node ",
+    "clean": "npx rimraf dist js ts/output ts/json ts/analyse ts/actions coverage"
   },
   "maintainers": [
     "Volker Sorge <Volker.Sorge@gmail.com> (http://www.cs.bham.ac.uk/~vxs)"
@@ -42,44 +42,41 @@
   ],
   "homepage": "https://github.com/speech-rule-engine/sre-tests",
   "dependencies": {
-    "@google-cloud/storage": "5.16.1",
-    "eslint-config-prettier": "^8.3.0",
+    "@google-cloud/storage": "6.10.1",
+    "@xmldom/xmldom": "^0.9.0-beta.8",
+    "eslint-config-prettier": "^8.8.0",
     "esm": "3.2.25",
-    "firebase-admin": "10.0.1",
-    "firebase-tools": "9.23.3",
-    "jest": "27.4.5",
-    "mathjax-full": "3.2.0",
-    "npm-check-updates": "12.0.5",
-    "wicked-good-xpath": "1.3.0",
-    "xmldom-sre": "0.1.31"
+    "firebase-admin": "^11.8.0",
+    "firebase-tools": "^12.2.1",
+    "jest": "29.5.0",
+    "mathjax-full": "3.2.2",
+    "npm-check-updates": "16.10.12",
+    "wicked-good-xpath": "1.3.0"
   },
   "devDependencies": {
-    "@babel/preset-env": "7.16.5",
-    "@types/jest": "27.0.3",
-    "@types/node": "17.0.1",
-    "@typescript-eslint/eslint-plugin": "5.7.0",
-    "@typescript-eslint/eslint-plugin-tslint": "5.7.0",
-    "@typescript-eslint/parser": "5.7.0",
-    "babel-jest": "27.4.5",
+    "@babel/preset-env": "7.22.4",
+    "@types/jest": "29.5.1",
+    "@types/node": "20.2.5",
+    "@typescript-eslint/eslint-plugin": "5.59.8",
+    "@typescript-eslint/parser": "5.59.8",
+    "babel-jest": "29.5.0",
     "babel-plugin-transform-class-properties": "6.24.1",
-    "eslint": "8.5.0",
-    "eslint-plugin-import": "2.25.3",
-    "eslint-plugin-jsdoc": "37.4.0",
-    "prettier": "^2.5.1",
-    "terser-webpack-plugin": "5.3.0",
-    "ts-loader": "9.2.6",
-    "tslint": "6.1.3",
-    "tslint-jsdoc-rules": "0.2.0",
-    "tslint-unix-formatter": "0.2.0",
-    "typescript": "4.5.4",
+    "eslint": "8.41.0",
+    "eslint-plugin-import": "2.27.5",
+    "eslint-plugin-jsdoc": "45.0.0",
+    "prettier": "^2.8.8",
+    "terser-webpack-plugin": "5.3.9",
+    "ts-loader": "9.4.3",
+    "typescript": "5.0.4",
     "typescript-tools": "0.3.1",
-    "webpack": "5.65.0",
-    "webpack-cli": "4.9.1"
+    "webpack": "5.84.1",
+    "webpack-cli": "5.1.1"
   },
   "files": [
     "LICENSE",
     "README.md",
     "lib/"
   ],
-  "main": "dist/sretests.js"
+  "main": "dist/sretests.js",
+  "type": "module"
 }
