@@ -1,2 +1,2 @@
-!function(t,e){"object"==typeof exports&&"object"==typeof module?module.exports=e():"function"==typeof define&&define.amd?define([],e):"object"==typeof exports?exports.Nemeth=e():t.Nemeth=e()}(this,(function(){return(()=>{"use strict";var t={647:(t,e)=>{Object.defineProperty(e,"__esModule",{value:!0}),e.NemethProjectUser=e.NemethProjectPath=e.NemethProjectDocuments=e.FeedbackStatus=e.Interaction=e.TestsCollection=e.NemethCollection=e.NemethProject=e.NemethUrl=void 0,e.NemethUrl="https://nemeth-project.firebaseio.com",e.NemethProject="nemeth-project",e.NemethCollection="nemeth",e.TestsCollection="tests",e.Interaction="InteractionStatus",e.FeedbackStatus="feedback",e.NemethProjectDocuments=e.NemethProject+"::documents",e.NemethProjectPath=e.NemethProject+"::path",e.NemethProjectUser=e.NemethProject+"::user"},416:function(t,e,n){var r=this&&this.__awaiter||function(t,e,n,r){return new(n||(n=Promise))((function(o,i){function a(t){try{c(r.next(t))}catch(t){i(t)}}function s(t){try{c(r.throw(t))}catch(t){i(t)}}function c(t){var e;t.done?o(t.value):(e=t.value,e instanceof n?e:new n((function(t){t(e)}))).then(a,s)}c((r=r.apply(t,e||[])).next())}))},o=this&&this.__generator||function(t,e){var n,r,o,i,a={label:0,sent:function(){if(1&o[0])throw o[1];return o[1]},trys:[],ops:[]};return i={next:s(0),throw:s(1),return:s(2)},"function"==typeof Symbol&&(i[Symbol.iterator]=function(){return this}),i;function s(i){return function(s){return function(i){if(n)throw new TypeError("Generator is already executing.");for(;a;)try{if(n=1,r&&(o=2&i[0]?r.return:i[0]?r.throw||((o=r.return)&&o.call(r),0):r.next)&&!(o=o.call(r,i[1])).done)return o;switch(r=0,o&&(i=[2&i[0],o.value]),i[0]){case 0:case 1:o=i;break;case 4:return a.label++,{value:i[1],done:!1};case 5:a.label++,r=i[1],i=[0];continue;case 7:i=a.ops.pop(),a.trys.pop();continue;default:if(!(o=a.trys,(o=o.length>0&&o[o.length-1])||6!==i[0]&&2!==i[0])){a=0;continue}if(3===i[0]&&(!o||i[1]>o[0]&&i[1]<o[3])){a.label=i[1];break}if(6===i[0]&&a.label<o[1]){a.label=o[1],o=i;break}if(o&&a.label<o[2]){a.label=o[2],a.ops.push(i);break}o[2]&&a.ops.pop(),a.trys.pop();continue}i=e.call(t,a)}catch(t){i=[6,t],r=0}finally{n=o=0}if(5&i[0])throw i[1];return{value:i[0]?i[1]:void 0,done:!0}}([i,s])}}},i=this&&this.__values||function(t){var e="function"==typeof Symbol&&Symbol.iterator,n=e&&t[e],r=0;if(n)return n.call(t);if(t&&"number"==typeof t.length)return{next:function(){return t&&r>=t.length&&(t=void 0),{value:t&&t[r++],done:!t}}};throw new TypeError(e?"Object is not iterable.":"Symbol.iterator is not defined.")};Object.defineProperty(e,"__esModule",{value:!0}),e.FireTest=void 0;var a=n(647),s=n(900),c=function(){function t(t,e,n,r,o){this.db=t,this.collection=e,this.doc=n,this.getTest=r,this.setTest=o,this.update=!0,this.countTests=0,this.preparedTests=[]}return t.prototype.getData=function(){return this._data},t.prototype.currentTest=function(){return this.preparedTests[this.countTests]},t.prototype.prepareTests=function(){return r(this,void 0,void 0,(function(){var t,e,n,r,a,c,u;return o(this,(function(o){switch(o.label){case 0:return t=this,[4,s.downloadData(this.db,this.collection,this.doc)];case 1:t._data=o.sent(),this.order=this._data.order,this.tests=this._data.tests;try{for(e=i(this.order),n=e.next();!n.done;n=e.next())r=n.value,(a=this.tests[r]).brf="",a.unicode="",this.preparedTests.push(a)}catch(t){c={error:t}}finally{try{n&&!n.done&&(u=e.return)&&u.call(e)}finally{if(c)throw c.error}}return[2]}}))}))},t.prototype.saveTest=function(t){return r(this,void 0,void 0,(function(){var e,n,r,c,u,l,f;return o(this,(function(o){e=this.currentTest(),n=null;try{for(r=i(Object.keys(t)),c=r.next();!c.done;c=r.next())u=c.value,e[u]!==t[u]&&(n=2),e[u]=t[u]}catch(t){l={error:t}}finally{try{c&&!c.done&&(f=r.return)&&f.call(r)}finally{if(l)throw l.error}}return 2===n&&(e[a.Interaction]=n,this.update&&s.updateData(this.db,this.collection,this.doc,e,["tests",e.name])),n||e[a.Interaction]||(e[a.Interaction]=1,this.update&&s.updateData(this.db,this.collection,this.doc,e[a.Interaction],["tests",e.name,a.Interaction])),[2]}))}))},t.prototype.saveFeedback=function(t){return r(this,void 0,void 0,(function(){var e;return o(this,(function(n){return e=this.currentTest(),s.updateData(this.db,this.collection,this.doc,t,["tests",e.name,a.FeedbackStatus]),[2]}))}))},t.prototype.cycleTests=function(t){return r(this,void 0,void 0,(function(){return o(this,(function(e){switch(e.label){case 0:return[4,this.saveTest(this.getTest())];case 1:return e.sent(),this.setTest(this.nextTest(t)),[2]}}))}))},t.prototype.cycleUnchangedTests=function(t){return r(this,void 0,void 0,(function(){return o(this,(function(e){switch(e.label){case 0:return[4,this.saveTest(this.getTest())];case 1:return e.sent(),this.setTest(this.jumpTest(t,(function(t){return 2!==t}))),[2]}}))}))},t.prototype.cycleNewTests=function(t){return r(this,void 0,void 0,(function(){return o(this,(function(e){switch(e.label){case 0:return[4,this.saveTest(this.getTest())];case 1:return e.sent(),this.setTest(this.jumpTest(t,(function(t){return 0===t}))),[2]}}))}))},t.prototype.goTest=function(t){return r(this,void 0,void 0,(function(){var e;return o(this,(function(n){switch(n.label){case 0:return[4,this.saveTest(this.getTest())];case 1:return n.sent(),-1===(e=this.order.indexOf(t))?[2]:(this.countTests=e,this.setTest(this.currentTest()),[2])}}))}))},t.prototype.nextTest=function(t){return this.countTests=this.countTests+(t?1:-1),this.countTests<0&&(this.countTests=this.preparedTests.length-1),this.countTests>=this.preparedTests.length&&(this.countTests=0),this.currentTest()},t.prototype.jumpTest=function(t,e){for(var n=this.countTests;!e(this.nextTest(t)[a.Interaction])&&this.countTests!==n;);return n!==this.countTests?this.currentTest():this.nextTest(t)},t}();e.FireTest=c},900:function(t,e,n){var r=this&&this.__awaiter||function(t,e,n,r){return new(n||(n=Promise))((function(o,i){function a(t){try{c(r.next(t))}catch(t){i(t)}}function s(t){try{c(r.throw(t))}catch(t){i(t)}}function c(t){var e;t.done?o(t.value):(e=t.value,e instanceof n?e:new n((function(t){t(e)}))).then(a,s)}c((r=r.apply(t,e||[])).next())}))},o=this&&this.__generator||function(t,e){var n,r,o,i,a={label:0,sent:function(){if(1&o[0])throw o[1];return o[1]},trys:[],ops:[]};return i={next:s(0),throw:s(1),return:s(2)},"function"==typeof Symbol&&(i[Symbol.iterator]=function(){return this}),i;function s(i){return function(s){return function(i){if(n)throw new TypeError("Generator is already executing.");for(;a;)try{if(n=1,r&&(o=2&i[0]?r.return:i[0]?r.throw||((o=r.return)&&o.call(r),0):r.next)&&!(o=o.call(r,i[1])).done)return o;switch(r=0,o&&(i=[2&i[0],o.value]),i[0]){case 0:case 1:o=i;break;case 4:return a.label++,{value:i[1],done:!1};case 5:a.label++,r=i[1],i=[0];continue;case 7:i=a.ops.pop(),a.trys.pop();continue;default:if(!(o=a.trys,(o=o.length>0&&o[o.length-1])||6!==i[0]&&2!==i[0])){a=0;continue}if(3===i[0]&&(!o||i[1]>o[0]&&i[1]<o[3])){a.label=i[1];break}if(6===i[0]&&a.label<o[1]){a.label=o[1],o=i;break}if(o&&a.label<o[2]){a.label=o[2],a.ops.push(i);break}o[2]&&a.ops.pop(),a.trys.pop();continue}i=e.call(t,a)}catch(t){i=[6,t],r=0}finally{n=o=0}if(5&i[0])throw i[1];return{value:i[0]?i[1]:void 0,done:!0}}([i,s])}}},i=this&&this.__values||function(t){var e="function"==typeof Symbol&&Symbol.iterator,n=e&&t[e],r=0;if(n)return n.call(t);if(t&&"number"==typeof t.length)return{next:function(){return t&&r>=t.length&&(t=void 0),{value:t&&t[r++],done:!t}}};throw new TypeError(e?"Object is not iterable.":"Symbol.iterator is not defined.")},a=this&&this.__read||function(t,e){var n="function"==typeof Symbol&&t[Symbol.iterator];if(!n)return t;var r,o,i=n.call(t),a=[];try{for(;(void 0===e||e-- >0)&&!(r=i.next()).done;)a.push(r.value)}catch(t){o={error:t}}finally{try{r&&!r.done&&(n=i.return)&&n.call(i)}finally{if(o)throw o.error}}return a};Object.defineProperty(e,"__esModule",{value:!0}),e.restoreField=e.updateField=e.updateCollection=e.getPaths=e.setPath=e.downloadData=e.updateData=e.uploadData=e.verbose=void 0;var s=n(647);function c(t){e.verbose&&console.log(t)}function u(t,e,n,i){return r(this,void 0,void 0,(function(){return o(this,(function(r){return t.collection(e).doc(n).set(i).then((function(){c("Data for document ".concat(n," successfully uploaded to collection ").concat(e))})).catch((function(t){c("Uploading data to document ".concat(n," in collection ").concat(e," failed with: ").concat(t))})),[2]}))}))}function l(t,e,n,i,a){return r(this,void 0,void 0,(function(){var r,s;return o(this,(function(o){return r={},s=a.join("."),r[s]=i,t.collection(e).doc(n).update(r).then((function(){c("Data successfully updated in collection ".concat(e," with path ").concat(n))})).catch((function(t){c("Updating data in collection ".concat(e," with path ").concat(n," failed with: ").concat(t))})),[2]}))}))}function f(t,e,n){return r(this,void 0,void 0,(function(){return o(this,(function(r){switch(r.label){case 0:return[4,t.collection(e).doc(n).get()];case 1:return[2,r.sent().data()]}}))}))}function d(t,e,n,i){return void 0===i&&(i=!0),r(this,void 0,void 0,(function(){var r,a,s;return o(this,(function(o){switch(o.label){case 0:return r=n.split("/")[0],[4,t.collection(e).doc(r).get()];case 1:return a=o.sent(),(s=a.data())||(s={}),s[n]=i,u(t,e,r,s),[2]}}))}))}function h(t,e,n){return r(this,void 0,void 0,(function(){var r;return o(this,(function(o){switch(o.label){case 0:return r=n.split("/")[0],[4,t.collection(e).doc(r).get()];case 1:return[2,o.sent().data()]}}))}))}function p(t){t[s.Interaction]=0}function v(t){t[s.FeedbackStatus]=0}e.verbose=!1,e.uploadData=u,e.updateData=l,e.downloadData=f,e.setPath=d,e.getPaths=h,e.updateCollection=function(t,e,n,s,c){return void 0===c&&(c=function(t,e){}),r(this,void 0,void 0,(function(){var r,y,b,m,w,g,T,x,S,k,_,B,j,N,L,E,C,P,U,M,A,I,O;return o(this,(function(o){switch(o.label){case 0:return r=[],[4,h(t,e,s)];case 1:return(y=o.sent())?[4,h(t,n,s)]:[2,r];case 2:b=o.sent(),o.label=3;case 3:o.trys.push([3,21,22,23]),m=i(Object.entries(y)),w=m.next(),o.label=4;case 4:return w.done?[3,20]:(g=a(w.value,2),T=g[0],x=a(g[1],2),S=x[0],k=x[1],c(T,Object.keys(y)),r.push({name:S,information:k,path:T}),b&&b[T]&&!T.match(/^nemeth\/rules/)?[3,19]:[4,f(t,e,T)]);case 5:return _=o.sent(),[4,f(t,n,T)];case 6:return(B=o.sent())?[3,9]:(function(t){var e,n;try{for(var r=i(Object.values(t.tests)),o=r.next();!o.done;o=r.next()){p(o.value)}}catch(t){e={error:t}}finally{try{o&&!o.done&&(n=r.return)&&n.call(r)}finally{if(e)throw e.error}}}(_),function(t){var e,n;try{for(var r=i(Object.values(t.tests)),o=r.next();!o.done;o=r.next()){v(o.value)}}catch(t){e={error:t}}finally{try{o&&!o.done&&(n=r.return)&&n.call(r)}finally{if(e)throw e.error}}}(_),[4,u(t,n,T,_)]);case 7:return o.sent(),[4,d(t,n,T)];case 8:return o.sent(),[3,19];case 9:j=B.tests,o.label=10;case 10:o.trys.push([10,15,16,17]),I=void 0,N=i(_.order),L=N.next(),o.label=11;case 11:return L.done?[3,14]:(E=L.value,j[E]?[3,13]:(p(C=_.tests[E]),v(C),[4,l(t,n,T,C,["tests",E])]));case 12:o.sent(),o.label=13;case 13:return L=N.next(),[3,11];case 14:return[3,17];case 15:return P=o.sent(),I={error:P},[3,17];case 16:try{L&&!L.done&&(O=N.return)&&O.call(N)}finally{if(I)throw I.error}return[7];case 17:return[4,l(t,n,T,_.order,["order"])];case 18:o.sent(),o.label=19;case 19:return w=m.next(),[3,4];case 20:return[3,23];case 21:return U=o.sent(),M={error:U},[3,23];case 22:try{w&&!w.done&&(A=m.return)&&A.call(m)}finally{if(M)throw M.error}return[7];case 23:return[2,r]}}))}))},e.updateField=function(t,e,n,a,s){return r(this,void 0,void 0,(function(){var r,c,u,d,h,p,v;return o(this,(function(o){switch(o.label){case 0:return[4,f(t,e,n)];case 1:r=o.sent(),o.label=2;case 2:o.trys.push([2,7,8,9]),c=i(r.order),u=c.next(),o.label=3;case 3:return u.done?[3,6]:(d=u.value,[4,l(t,e,n,s,["tests",d,a])]);case 4:o.sent(),o.label=5;case 5:return u=c.next(),[3,3];case 6:return[3,9];case 7:return h=o.sent(),p={error:h},[3,9];case 8:try{u&&!u.done&&(v=c.return)&&v.call(c)}finally{if(p)throw p.error}return[7];case 9:return[2]}}))}))},e.restoreField=function(t,e,n,a,s){return r(this,void 0,void 0,(function(){var r,c,u,d,h,p,v,y,b,m;return o(this,(function(o){switch(o.label){case 0:return[4,f(t,e,a)];case 1:return r=o.sent(),[4,f(t,n,a)];case 2:if(!(c=o.sent()))return[2];o.label=3;case 3:o.trys.push([3,8,9,10]),u=i(c.order),d=u.next(),o.label=4;case 4:return d.done?[3,7]:(h=d.value,(p=r.tests[h])?void 0===(v=p[s])?[3,6]:[4,l(t,n,a,v,["tests",h,s])]:[3,6]);case 5:o.sent(),o.label=6;case 6:return d=u.next(),[3,4];case 7:return[3,10];case 8:return y=o.sent(),b={error:y},[3,10];case 9:try{d&&!d.done&&(m=u.return)&&m.call(u)}finally{if(b)throw b.error}return[7];case 10:return[2]}}))}))}},986:function(t,e){var n=this&&this.__awaiter||function(t,e,n,r){return new(n||(n=Promise))((function(o,i){function a(t){try{c(r.next(t))}catch(t){i(t)}}function s(t){try{c(r.throw(t))}catch(t){i(t)}}function c(t){var e;t.done?o(t.value):(e=t.value,e instanceof n?e:new n((function(t){t(e)}))).then(a,s)}c((r=r.apply(t,e||[])).next())}))},r=this&&this.__generator||function(t,e){var n,r,o,i,a={label:0,sent:function(){if(1&o[0])throw o[1];return o[1]},trys:[],ops:[]};return i={next:s(0),throw:s(1),return:s(2)},"function"==typeof Symbol&&(i[Symbol.iterator]=function(){return this}),i;function s(i){return function(s){return function(i){if(n)throw new TypeError("Generator is already executing.");for(;a;)try{if(n=1,r&&(o=2&i[0]?r.return:i[0]?r.throw||((o=r.return)&&o.call(r),0):r.next)&&!(o=o.call(r,i[1])).done)return o;switch(r=0,o&&(i=[2&i[0],o.value]),i[0]){case 0:case 1:o=i;break;case 4:return a.label++,{value:i[1],done:!1};case 5:a.label++,r=i[1],i=[0];continue;case 7:i=a.ops.pop(),a.trys.pop();continue;default:if(!(o=a.trys,(o=o.length>0&&o[o.length-1])||6!==i[0]&&2!==i[0])){a=0;continue}if(3===i[0]&&(!o||i[1]>o[0]&&i[1]<o[3])){a.label=i[1];break}if(6===i[0]&&a.label<o[1]){a.label=o[1],o=i;break}if(o&&a.label<o[2]){a.label=o[2],a.ops.push(i);break}o[2]&&a.ops.pop(),a.trys.pop();continue}i=e.call(t,a)}catch(t){i=[6,t],r=0}finally{n=o=0}if(5&i[0])throw i[1];return{value:i[0]?i[1]:void 0,done:!0}}([i,s])}}},o=this&&this.__values||function(t){var e="function"==typeof Symbol&&Symbol.iterator,n=e&&t[e],r=0;if(n)return n.call(t);if(t&&"number"==typeof t.length)return{next:function(){return t&&r>=t.length&&(t=void 0),{value:t&&t[r++],done:!t}}};throw new TypeError(e?"Object is not iterable.":"Symbol.iterator is not defined.")};Object.defineProperty(e,"__esModule",{value:!0}),e.harvest=e.init=e.updateAccess=e.access=e.backwardBtn=e.forwardBtn=void 0;var i=function(){function t(e){var n,r;this.direction=e,this.span=t.createElement("btns"),this.f=t.createElement("btn f",this.direction+" element"),this.ff=t.createElement("btn ff",this.direction+" unchanged element"),this.fff=t.createElement("btn fff",this.direction+" unviewed element"),this.span.id=e;var i="next"===e?[this.f,this.ff,this.fff]:[this.fff,this.ff,this.f];try{for(var a=o(i),s=a.next();!s.done;s=a.next()){var c=s.value;c.classList.add(e),c.setAttribute("tabindex","0"),this.span.appendChild(c)}}catch(t){n={error:t}}finally{try{s&&!s.done&&(r=a.return)&&r.call(a)}finally{if(n)throw n.error}}this.hide()}return t.createElement=function(t,e){void 0===e&&(e="");var n=document.createElement("span");return n.className=t,e&&(n.title=e,n.setAttribute("aria-label",e)),n},t.prototype.show=function(){this.f.classList.remove("hidden"),this.ff.classList.remove("hidden"),this.fff.classList.remove("hidden")},t.prototype.hide=function(){this.f.classList.add("hidden"),this.ff.classList.add("hidden"),this.fff.classList.add("hidden")},t}(),a=function(){function t(){this.anchor=i.createElement("btn center","access"),this.select=document.createElement("select"),this.select.id="access",this.select.classList.add("access"),this.anchor.appendChild(this.select)}return t.prototype.init=function(t){var e,n,r=this;this.test=t,this.select.addEventListener("change",(function(t){r.test.goTest(t.target.value)}));try{for(var i=o(this.test.order),a=i.next();!a.done;a=i.next()){var s=a.value,c=document.createElement("option");c.setAttribute("value",s),c.classList.add("access"),c.textContent=s,this.select.appendChild(c)}}catch(t){e={error:t}}finally{try{a&&!a.done&&(n=i.return)&&n.call(i)}finally{if(e)throw e.error}}this.update()},t.prototype.update=function(){this.select.value=this.test.currentTest().name},t}();function s(t,e){t.addEventListener("click",e),t.addEventListener("keyup",(function(t){"Enter"===t.key&&e(t)}))}function c(t){var o=this;s(e.backwardBtn.f,(function(){return n(o,void 0,void 0,(function(){return r(this,(function(e){switch(e.label){case 0:return[4,t.cycleTests(!1)];case 1:return[2,e.sent()]}}))}))})),s(e.backwardBtn.ff,(function(){return n(o,void 0,void 0,(function(){return r(this,(function(e){switch(e.label){case 0:return[4,t.cycleUnchangedTests(!1)];case 1:return[2,e.sent()]}}))}))})),s(e.backwardBtn.fff,(function(){return n(o,void 0,void 0,(function(){return r(this,(function(e){switch(e.label){case 0:return[4,t.cycleNewTests(!1)];case 1:return[2,e.sent()]}}))}))})),s(e.forwardBtn.f,(function(){return n(o,void 0,void 0,(function(){return r(this,(function(e){switch(e.label){case 0:return[4,t.cycleTests(!0)];case 1:return[2,e.sent()]}}))}))})),s(e.forwardBtn.ff,(function(){return n(o,void 0,void 0,(function(){return r(this,(function(e){switch(e.label){case 0:return[4,t.cycleUnchangedTests(!0)];case 1:return[2,e.sent()]}}))}))})),s(e.forwardBtn.fff,(function(){return n(o,void 0,void 0,(function(){return r(this,(function(e){switch(e.label){case 0:return[4,t.cycleNewTests(!0)];case 1:return[2,e.sent()]}}))}))}))}e.forwardBtn=new i("next"),e.backwardBtn=new i("previous"),e.access=new a,e.updateAccess=function(){e.access.update()},e.init=function(t){c(t);var n=document.createElement("div");n.classList.add("btns"),document.body.appendChild(n),n.appendChild(e.backwardBtn.span),e.backwardBtn.show(),e.access.init(t),n.appendChild(e.access.anchor);var r=i.createElement("btn center","selection");r.textContent="Selection",r.setAttribute("tabindex","0"),s(r,(function(){t.saveTest(t.getTest()),window.location.assign("selection.html")})),n.appendChild(r),n.appendChild(e.forwardBtn.span),e.forwardBtn.show()},e.harvest=function(t,n){void 0===n&&(n=function(){}),c(t);var r=document.createElement("div");r.classList.add("btns"),document.body.appendChild(r),r.appendChild(e.backwardBtn.span),e.backwardBtn.show(),e.access.init(t),r.appendChild(e.access.anchor);var o=i.createElement("btn center","selection");o.textContent="Save",o.setAttribute("tabindex","0"),s(o,(function(){t.saveTest(t.getTest()),n()})),r.appendChild(o),r.appendChild(e.forwardBtn.span),e.forwardBtn.show()}},362:function(t,e,n){var r=this&&this.__awaiter||function(t,e,n,r){return new(n||(n=Promise))((function(o,i){function a(t){try{c(r.next(t))}catch(t){i(t)}}function s(t){try{c(r.throw(t))}catch(t){i(t)}}function c(t){var e;t.done?o(t.value):(e=t.value,e instanceof n?e:new n((function(t){t(e)}))).then(a,s)}c((r=r.apply(t,e||[])).next())}))},o=this&&this.__generator||function(t,e){var n,r,o,i,a={label:0,sent:function(){if(1&o[0])throw o[1];return o[1]},trys:[],ops:[]};return i={next:s(0),throw:s(1),return:s(2)},"function"==typeof Symbol&&(i[Symbol.iterator]=function(){return this}),i;function s(i){return function(s){return function(i){if(n)throw new TypeError("Generator is already executing.");for(;a;)try{if(n=1,r&&(o=2&i[0]?r.return:i[0]?r.throw||((o=r.return)&&o.call(r),0):r.next)&&!(o=o.call(r,i[1])).done)return o;switch(r=0,o&&(i=[2&i[0],o.value]),i[0]){case 0:case 1:o=i;break;case 4:return a.label++,{value:i[1],done:!1};case 5:a.label++,r=i[1],i=[0];continue;case 7:i=a.ops.pop(),a.trys.pop();continue;default:if(!(o=a.trys,(o=o.length>0&&o[o.length-1])||6!==i[0]&&2!==i[0])){a=0;continue}if(3===i[0]&&(!o||i[1]>o[0]&&i[1]<o[3])){a.label=i[1];break}if(6===i[0]&&a.label<o[1]){a.label=o[1],o=i;break}if(o&&a.label<o[2]){a.label=o[2],a.ops.push(i);break}o[2]&&a.ops.pop(),a.trys.pop();continue}i=e.call(t,a)}catch(t){i=[6,t],r=0}finally{n=o=0}if(5&i[0])throw i[1];return{value:i[0]?i[1]:void 0,done:!0}}([i,s])}}},i=this&&this.__values||function(t){var e="function"==typeof Symbol&&Symbol.iterator,n=e&&t[e],r=0;if(n)return n.call(t);if(t&&"number"==typeof t.length)return{next:function(){return t&&r>=t.length&&(t=void 0),{value:t&&t[r++],done:!t}}};throw new TypeError(e?"Object is not iterable.":"Symbol.iterator is not defined.")},a=this&&this.__read||function(t,e){var n="function"==typeof Symbol&&t[Symbol.iterator];if(!n)return t;var r,o,i=n.call(t),a=[];try{for(;(void 0===e||e-- >0)&&!(r=i.next()).done;)a.push(r.value)}catch(t){o={error:t}}finally{try{r&&!r.done&&(n=i.return)&&n.call(i)}finally{if(o)throw o.error}}return a},s=this&&this.__spreadArray||function(t,e,n){if(n||2===arguments.length)for(var r,o=0,i=e.length;o<i;o++)!r&&o in e||(r||(r=Array.prototype.slice.call(e,0,o)),r[o]=e[o]);return t.concat(r||Array.prototype.slice.call(e))};Object.defineProperty(e,"__esModule",{value:!0}),e.generateMath=e.initHarvest=e.harvest=e.changeFeedback=e.changeFormat=e.generatem=e.generate=e.init=e.fireTest=void 0;var c=n(647),u=n(416),l=n(532),f=n(986),d=n(243),h=new Map([["NABT",new l.Nabt2Unicode],["BLDT",new l.Bldt2Unicode],["NABT-TABLE",new l.Nabt2UnicodeTable],["BLDT-TABLE",new l.Bldt2UnicodeTable],["SDF/JKL",new l.Ascii2Braille("","")],["1-6",new l.Numeric2Braille("","")]]),p=new Map([["NABT",new l.Unicode2Nabt],["BLDT",new l.Unicode2Bldt],["SDF/JKL",new l.Braille2Ascii("","")],["1-6",new l.Braille2Numeric("","")]]),v="NABT";function y(){return h.get(v.toUpperCase())}var b={};e.fireTest=null;var m="";function w(t){var e;b.mathname.innerHTML=t.name,b.mathexpression.innerHTML=t.input?'<math display="block">'.concat(t.input,"</math>"):t.tex?"\\[".concat(t.tex,"\\]"):"".concat(t.name),b.braille.innerHTML=t.expected,b.input.value=(console.log(8),console.log(v.toUpperCase()),p.get(v.toUpperCase())).via(t.expected),b.input.focus(),function(t){var e,n,r;if(b.refname.innerHTML="",b.references.innerHTML="",!t)return;var o=Object.keys(t).sort((function(t,e){var n=parseInt(t.split("-")[1],10),r=parseInt(e.split("-")[1],10);return n<r?-1:n>r?1:0})),c=o.length;if(!c)return;b.refname.innerHTML="References:";var u=[],l=c>10?T:g;try{for(var f=i(o),d=f.next();!d.done;d=f.next()){var h=d.value,p=t[h];u.push(l(h,p))}}catch(t){e={error:t}}finally{try{d&&!d.done&&(n=f.return)&&n.call(f)}finally{if(e)throw e.error}}if(c<=10)return void(r=b.references).append.apply(r,s([],a(u),!1));var v=document.createElement("select");v.setAttribute("onChange",'window.open(this.value, "_blank")'),v.classList.add("access");var y=document.createElement("option");y.setAttribute("disabled",""),y.setAttribute("selected",""),y.textContent="Select:",v.appendChild(y),v.append.apply(v,s([],a(u),!1)),b.references.appendChild(v)}(t.reference),x(t[c.Interaction]),void 0!==(e=t[c.FeedbackStatus])&&(b.feedback.value=e.toString()),MathJax.typeset&&MathJax.typeset(),(0,f.updateAccess)()}function g(t,e){var n=document.createElement("a");return n.classList.add("reflink"),n.setAttribute("href",e),n.setAttribute("target","_blank"),n.textContent=t,n}function T(t,e){var n=document.createElement("option");return n.classList.add("access"),n.textContent=t,n.value=e,n}function x(t){if(void 0!==t&&b.statuscolor&&b.statusvalue)switch(t){case 0:b.statuscolor.className="green",b.statusvalue.innerHTML="New";break;case 1:b.statuscolor.className="yellow",b.statusvalue.innerHTML="Viewed";break;case 2:b.statuscolor.className="red",b.statusvalue.innerHTML="Changed";break;default:b.statuscolor.className="",b.statusvalue.innerHTML=""}}function S(){return{expected:b.braille.innerHTML}}function k(t,n){firebase?function(t,n){r(this,void 0,void 0,(function(){var r;return o(this,(function(o){switch(o.label){case 0:return r=firebase.app().firestore(),e.fireTest=new u.FireTest(r,t,n,S,w),[4,e.fireTest.prepareTests()];case 1:return o.sent(),_(),(0,f.init)(e.fireTest),e.fireTest.setTest(e.fireTest.currentTest()),[2]}}))}))}(t,n):setTimeout(k,100)}function _(){for(var t=document.evaluate("//*[@id]",document.body),e=t.iterateNext();e;e=t.iterateNext())b[e.id]=e}function B(){var t=b.input;m!==t.value&&("SDF/JKL"===v.toUpperCase()&&function(){if(j)return;j=!0,setTimeout((function(){j=!1,N();var t=b.input;","!==t.value[t.value.length-1]&&(t.value=t.value+",")}),100)}(),N())}e.init=function(){var t=d.getStorage(c.NemethProjectPath),e=d.getStorage(c.NemethProjectUser);t&&e&&k(e,t)},e.generate=B;var j=!1;function N(){var t=b.input,e=t.selectionStart;b.braille.innerHTML="",b.error.innerHTML="";var n=a(function(t){var e=a(y().cleanInput(t),2),n=e[0],r=e[1];return[n,y().via(n),r]}(t.value),3),r=n[0],o=n[1],i=n[2];i&&(b.error.innerHTML="Unknown "+(i.length>1?"elements ":"element ")+i),r!==m&&x(2),t.value=r,m=r,o=o.split("\n").join("<br>"),b.braille.innerHTML=o,t.selectionEnd=e-i.length}function L(){b.mathinput.value="",b.input.value="",P(),B()}function E(t,n){return r(this,void 0,void 0,(function(){var r;return o(this,(function(o){switch(o.label){case 0:return r=firebase.app().firestore(),e.fireTest=new u.FireTest(r,t,n,S,w),[4,e.fireTest.prepareTests()];case 1:return o.sent(),_(),(0,f.harvest)(e.fireTest,L),console.log(n),console.log(t),[2]}}))}))}e.generatem=function(){setTimeout(B,100)},e.changeFormat=function(){e.fireTest.saveTest(e.fireTest.getTest()),v=b.format.value,e.fireTest.setTest(e.fireTest.currentTest())},e.changeFeedback=function(){var t=parseInt(b.feedback.value,10);e.fireTest.currentTest()[c.FeedbackStatus]=t,e.fireTest.saveFeedback(t)},e.harvest=function(){var t=d.getStorage(c.NemethProjectPath),e=d.getStorage(c.NemethProjectUser);t&&e&&E(e,t)},e.initHarvest=E;var C="";function P(){var t,e=b.mathinput;C!==e.value&&(C=e.value,t=e.value,b.mathexpression.innerHTML="\\[".concat(t,"\\]"),MathJax.typeset&&MathJax.typeset())}e.generateMath=P},243:(t,e)=>{Object.defineProperty(e,"__esModule",{value:!0}),e.setStorage=e.getStorage=e.getFirebase=void 0,e.getFirebase=function(){return firebase},e.getStorage=function(t){return localStorage.getItem(t)},e.setStorage=function(t,e){localStorage.setItem(t,e)}},139:function(t,e,n){var r=this&&this.__values||function(t){var e="function"==typeof Symbol&&Symbol.iterator,n=e&&t[e],r=0;if(n)return n.call(t);if(t&&"number"==typeof t.length)return{next:function(){return t&&r>=t.length&&(t=void 0),{value:t&&t[r++],done:!t}}};throw new TypeError(e?"Object is not iterable.":"Symbol.iterator is not defined.")};Object.defineProperty(e,"__esModule",{value:!0}),e.getUid=e.addDocuments=e.addSection=void 0;var o=n(647),i=n(243);function a(t,e){var n=i.getStorage(o.NemethProjectDocuments);n&&function(t,e){var n,a;t.sort((function(t,e){return(""+t.name).localeCompare(e.name)}));var c=function(t){var n=document.createElement("tr");n.setAttribute("tabindex","0"),n.classList.add("selection"),n.innerHTML='<td class="selection">'.concat(t.name,"</td>")+"<td>".concat(t.information,"</td><td>").concat(t.path,"</td>"),n.addEventListener("click",(function(){i.setStorage(o.NemethProjectPath,t.path),i.setStorage(o.NemethProjectUser,s()),window.location.assign("brf2nemeth.html")})),n.addEventListener("keyup",(function(e){"Enter"===e.key&&(i.setStorage(o.NemethProjectPath,t.path),i.setStorage(o.NemethProjectUser,s()),window.location.assign("brf2nemeth.html"))})),e.appendChild(n)};try{for(var u=r(t),l=u.next();!l.done;l=u.next()){c(l.value)}}catch(t){n={error:t}}finally{try{l&&!l.done&&(a=u.return)&&a.call(u)}finally{if(n)throw n.error}}}(JSON.parse(n).filter((function(t){return t.path.match(e)})),t)}function s(){return i.getFirebase().auth().getUid()}e.addSection=function(t,e,n,r){n="Select ".concat(n," to work on:");var o=document.createElement("hr");o.setAttribute("style","width:100%;border-width:0;background-color:#B1DF84;color:#B1DF84;height:6px");var i=document.createElement("div");i.classList.add("converter"),i.id=t+"table";var s=document.createElement("div");s.classList.add("section"),s.textContent=n;var c=function(t,e){var n=document.createElement("table");n.id=t;var r=document.createElement("tr");return r.classList.add("heading"),r.innerHTML="<th>Name</th><th>Information</th><th>Path</th>",n.appendChild(r),a(n,e),n}(t,e);i.append(s,c),r.append(o,i)},e.addDocuments=a,e.getUid=s},277:function(t,e,n){var r=this&&this.__awaiter||function(t,e,n,r){return new(n||(n=Promise))((function(o,i){function a(t){try{c(r.next(t))}catch(t){i(t)}}function s(t){try{c(r.throw(t))}catch(t){i(t)}}function c(t){var e;t.done?o(t.value):(e=t.value,e instanceof n?e:new n((function(t){t(e)}))).then(a,s)}c((r=r.apply(t,e||[])).next())}))},o=this&&this.__generator||function(t,e){var n,r,o,i,a={label:0,sent:function(){if(1&o[0])throw o[1];return o[1]},trys:[],ops:[]};return i={next:s(0),throw:s(1),return:s(2)},"function"==typeof Symbol&&(i[Symbol.iterator]=function(){return this}),i;function s(i){return function(s){return function(i){if(n)throw new TypeError("Generator is already executing.");for(;a;)try{if(n=1,r&&(o=2&i[0]?r.return:i[0]?r.throw||((o=r.return)&&o.call(r),0):r.next)&&!(o=o.call(r,i[1])).done)return o;switch(r=0,o&&(i=[2&i[0],o.value]),i[0]){case 0:case 1:o=i;break;case 4:return a.label++,{value:i[1],done:!1};case 5:a.label++,r=i[1],i=[0];continue;case 7:i=a.ops.pop(),a.trys.pop();continue;default:if(!(o=a.trys,(o=o.length>0&&o[o.length-1])||6!==i[0]&&2!==i[0])){a=0;continue}if(3===i[0]&&(!o||i[1]>o[0]&&i[1]<o[3])){a.label=i[1];break}if(6===i[0]&&a.label<o[1]){a.label=o[1],o=i;break}if(o&&a.label<o[2]){a.label=o[2],a.ops.push(i);break}o[2]&&a.ops.pop(),a.trys.pop();continue}i=e.call(t,a)}catch(t){i=[6,t],r=0}finally{n=o=0}if(5&i[0])throw i[1];return{value:i[0]?i[1]:void 0,done:!0}}([i,s])}}};Object.defineProperty(e,"__esModule",{value:!0}),e.init=void 0;var i=n(647),a=n(900),s=n(243);function c(t,e,n,c){return r(this,void 0,void 0,(function(){var r,u,l;return o(this,(function(o){switch(o.label){case 0:return r=0,u=document.createElement("div"),document.body.appendChild(u),[4,a.updateCollection(t,e,n,c,(function(t,e){var n=document.createElement("span");n.textContent=".",document.body.insertBefore(n,u),u.textContent="File (".concat(++r,"/").concat(e.length,")")}))];case 1:return l=o.sent(),s.setStorage(i.NemethProjectDocuments,JSON.stringify(l)),[2]}}))}))}var u={callbacks:{signInSuccessWithAuthResult:function(t){return function(t){r(this,void 0,void 0,(function(){var e;return o(this,(function(n){switch(n.label){case 0:return(e=document.createElement("span")).textContent="Updating.",document.body.appendChild(e),[4,c(s.getFirebase().firestore(),i.TestsCollection,t.user.uid,i.NemethCollection)];case 1:return n.sent(),window.location.assign("selection.html"),[2]}}))}))}(t),!1},uiShown:function(){document.getElementById("loader").style.display="none"}},signInSuccessUrl:"brf2nemeth.html",signInOptions:[s.getFirebase().auth.EmailAuthProvider.PROVIDER_ID],tosUrl:"privacy.html",privacyPolicyUrl:"privacy.html"};e.init=function(t,e,n){void 0===n&&(n=u),new t.auth.AuthUI(s.getFirebase().auth()).start(e,n)}},967:(t,e)=>{function n(t){document.getElementById("user-signed-in").style.display="inline",document.getElementById("name").textContent=t.displayName,document.getElementById("email").textContent=t.email}function r(){document.getElementById("user-signed-in").style.display="none",window.location.replace(".")}function o(){window.confirm("Are you sure you want to delete your account?")&&firebase.auth().currentUser.delete().catch((function(t){"auth/requires-recent-login"===t.code&&firebase.auth().signOut().then((function(){setTimeout((function(){alert("Please sign in again to delete your account.")}),1)}))}))}function i(){var t=document.createElement("span");t.innerHTML='<span id="user-signed-in" class="hidden"><span id="user-info"><span id="name"></span>(<span id="email"></span>)<div><span><button id="sign-out">Sign Out</button></span><span><button id="delete-account">Delete account</button></span></div></span></span>',document.body.appendChild(t),document.getElementById("sign-out").addEventListener("click",(function(){firebase.auth().signOut()})),document.getElementById("delete-account").addEventListener("click",(function(){o()}))}Object.defineProperty(e,"__esModule",{value:!0}),e.init=e.addSpan=e.deleteAccount=e.handleSignedOutUser=e.handleSignedInUser=void 0,e.handleSignedInUser=n,e.handleSignedOutUser=r,e.deleteAccount=o,e.addSpan=i,e.init=function(){i(),firebase.auth().onAuthStateChanged((function(t){t?n(t):r()}))}},532:function(t,e,n){var r,o=this&&this.__extends||(r=function(t,e){return r=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(t,e){t.__proto__=e}||function(t,e){for(var n in e)Object.prototype.hasOwnProperty.call(e,n)&&(t[n]=e[n])},r(t,e)},function(t,e){if("function"!=typeof e&&null!==e)throw new TypeError("Class extends value "+String(e)+" is not a constructor or null");function n(){this.constructor=t}r(t,e),t.prototype=null===e?Object.create(e):(n.prototype=e.prototype,new n)}),i=this&&this.__values||function(t){var e="function"==typeof Symbol&&Symbol.iterator,n=e&&t[e],r=0;if(n)return n.call(t);if(t&&"number"==typeof t.length)return{next:function(){return t&&r>=t.length&&(t=void 0),{value:t&&t[r++],done:!t}}};throw new TypeError(e?"Object is not iterable.":"Symbol.iterator is not defined.")};Object.defineProperty(e,"__esModule",{value:!0}),e.Braille2Ascii=e.Braille2Numeric=e.Ascii2Braille=e.Numeric2Braille=e.Unicode2Bldt=e.Unicode2Nabt=e.Bldt2UnicodeTable=e.Nabt2UnicodeTable=e.Bldt2Unicode=e.Nabt2Unicode=void 0;var a=function(t){function e(e,n){var r=t.call(this,e,n)||this;return r.translate=new Map,r.setupMap(),r}return o(e,t),e.getFormat=function(t){return t.toUpperCase().match(/^BLDT/)?e.format.BLDT:e.format.NABT},e.prototype.via=function(t){var e,n,r="";try{for(var o=i(t.split("")),a=o.next();!a.done;a=o.next()){var s=a.value,c=this.translate.get(s.toLowerCase());if(!c)throw new Error("Illegal input character: "+s);r+=c}}catch(t){e={error:t}}finally{try{a&&!a.done&&(n=o.return)&&n.call(o)}finally{if(e)throw e.error}}return r},e.prototype.cleanInput=function(t){var e,n,r=[],o=[];try{for(var a=i(t.toLowerCase().split("")),s=a.next();!s.done;s=a.next()){var c=s.value;void 0!==this.translate.get(c)?r.push(c):o.push(c)}}catch(t){e={error:t}}finally{try{s&&!s.done&&(n=a.return)&&n.call(a)}finally{if(e)throw e.error}}return[r.join(""),o.join(", ")]},e.format={NABT:" a1b'k2l@cif/msp\"e3h9o6r^djg>ntq,*5<-u8v.%[$+x!&;:4\\0z7(_?w]#y)=",BLDT:" a1b'k2l`cif/msp\"e3h9o6r~djg>ntq,*5<-u8v.%{$+x!&;:4|0z7(_?w}#y)="},e}(n(110).AbstractTransformer),s=function(t){function e(){var e=t.call(this,"expected","")||this;return e.dst=e.kind(),e}return o(e,t),e.prototype.setupMap=function(){var t,e,n=0;try{for(var r=i(a.getFormat(this.kind()).split("")),o=r.next();!o.done;o=r.next()){var s=o.value;this.translate.set(String.fromCodePoint(10240+n++),s)}}catch(e){t={error:e}}finally{try{o&&!o.done&&(e=r.return)&&e.call(r)}finally{if(t)throw t.error}}},e}(a),c=function(t){function e(){var e=t.call(this,"","expected")||this;return e.src=e.kind(),e}return o(e,t),e.prototype.setupMap=function(){var t,e,n=0;try{for(var r=i(a.getFormat(this.kind()).split("")),o=r.next();!o.done;o=r.next()){var s=o.value;this.translate.set(s,String.fromCodePoint(10240+n++))}}catch(e){t={error:e}}finally{try{o&&!o.done&&(e=r.return)&&e.call(r)}finally{if(t)throw t.error}}},e}(a),u=function(t){function e(){return null!==t&&t.apply(this,arguments)||this}return o(e,t),e.prototype.kind=function(){return"NABT"},e}(c);e.Nabt2Unicode=u;var l=function(t){function e(){return null!==t&&t.apply(this,arguments)||this}return o(e,t),e.prototype.kind=function(){return"BLDT"},e}(c);e.Bldt2Unicode=l;var f=function(t){function e(){return null!==t&&t.apply(this,arguments)||this}return o(e,t),e.prototype.kind=function(){return"NABT-TABLE"},e.prototype.via=function(e){var n,r,o=[];try{for(var a=i(e.split("\n")),s=a.next();!s.done;s=a.next()){var c=s.value;o.push(t.prototype.via.call(this,c))}}catch(t){n={error:t}}finally{try{s&&!s.done&&(r=a.return)&&r.call(a)}finally{if(n)throw n.error}}return console.log(o.join("\n")),o.join("\n")},e.prototype.cleanInput=function(t){var e,n,r,o,a=[],s=[];try{for(var c=i(t.split("\n")),u=c.next();!u.done;u=c.next()){var l=u.value;try{for(var f=(r=void 0,i(l.toLowerCase().split(""))),d=f.next();!d.done;d=f.next()){var h=d.value;void 0!==this.translate.get(h)?a.push(h):s.push(h)}}catch(t){r={error:t}}finally{try{d&&!d.done&&(o=f.return)&&o.call(f)}finally{if(r)throw r.error}}a.push("\n")}}catch(t){e={error:t}}finally{try{u&&!u.done&&(n=c.return)&&n.call(c)}finally{if(e)throw e.error}}return a.pop(),[a.join(""),s.join(", ")]},e}(u);e.Nabt2UnicodeTable=f;var d=function(t){function e(){return null!==t&&t.apply(this,arguments)||this}return o(e,t),e.prototype.kind=function(){return"BLDT-TABLE"},e.prototype.via=function(e){var n,r,o=[];try{for(var a=i(e.split("\n")),s=a.next();!s.done;s=a.next()){var c=s.value;o.push(t.prototype.via.call(this,c))}}catch(t){n={error:t}}finally{try{s&&!s.done&&(r=a.return)&&r.call(a)}finally{if(n)throw n.error}}return console.log(o.join("\n")),o.join("\n")},e.prototype.cleanInput=function(t){var e,n,r,o,a=[],s=[];try{for(var c=i(t.split("\n")),u=c.next();!u.done;u=c.next()){var l=u.value;try{for(var f=(r=void 0,i(l.toLowerCase().split(""))),d=f.next();!d.done;d=f.next()){var h=d.value;void 0!==this.translate.get(h)?a.push(h):s.push(h)}}catch(t){r={error:t}}finally{try{d&&!d.done&&(o=f.return)&&o.call(f)}finally{if(r)throw r.error}}a.push("\n")}}catch(t){e={error:t}}finally{try{u&&!u.done&&(n=c.return)&&n.call(c)}finally{if(e)throw e.error}}return a.pop(),[a.join(""),s.join(", ")]},e}(l);e.Bldt2UnicodeTable=d;var h=function(t){function e(){return null!==t&&t.apply(this,arguments)||this}return o(e,t),e.prototype.kind=function(){return"NABT"},e}(s);e.Unicode2Nabt=h;var p=function(t){function e(){return null!==t&&t.apply(this,arguments)||this}return o(e,t),e.prototype.kind=function(){return"BLDT"},e}(s);e.Unicode2Bldt=p;var v=function(t){function e(){return null!==t&&t.apply(this,arguments)||this}return o(e,t),e.prototype.via=function(t){return t.split(",").filter((function(t){return t})).map(this.toBraille.bind(this)).join("")},e}(a),y=function(t){function e(){return null!==t&&t.apply(this,arguments)||this}return o(e,t),e.prototype.via=function(t){return t.split("").map(this.fromBraille.bind(this)).join(",")},e}(a),b=function(t){function e(){return null!==t&&t.apply(this,arguments)||this}return o(e,t),e.prototype.kind=function(){return"1-6"},e.numToBraille=function(t){var e,n,r={},o=0;try{for(var a=i(t),s=a.next();!s.done;s=a.next()){var c=s.value;!r[c]&&c>0&&c<9&&(o+=Math.pow(2,c-1)),r[c]=!0}}catch(t){e={error:t}}finally{try{s&&!s.done&&(n=a.return)&&n.call(a)}finally{if(e)throw e.error}}var u=parseInt("2800",16);return String.fromCodePoint(u+o)},e.prototype.toBraille=function(t){return e.numToBraille(t.split("").map((function(t){return parseInt(t,10)})).filter((function(t){return!isNaN(t)})))},e.prototype.setupMap=function(){for(var t=0;t<=8;t++)this.translate.set(t.toString(),"T");this.translate.set(",","T")},e}(v);e.Numeric2Braille=b;var m=function(t){function e(){return null!==t&&t.apply(this,arguments)||this}return o(e,t),e.prototype.kind=function(){return"sdf/jkl"},e.prototype.toBraille=function(t){var e,n,r=[];try{for(var o=i(t.split("")),a=o.next();!a.done;a=o.next()){var s=a.value,c=parseInt(this.translate.get(s),10);isNaN(c)||r.push(c)}}catch(t){e={error:t}}finally{try{a&&!a.done&&(n=o.return)&&n.call(o)}finally{if(e)throw e.error}}return b.numToBraille(r)},e.prototype.setupMap=function(){this.translate=new Map([[",","T"],[" ","0"],["f","1"],["d","2"],["s","3"],["j","4"],["k","5"],["l","6"]])},e}(b);e.Ascii2Braille=m;var w=function(t){function e(){return null!==t&&t.apply(this,arguments)||this}return o(e,t),e.prototype.kind=function(){return"1-6"},e.unicodeToNumber=function(t){var e=t.codePointAt(0)-10240;if(!e)return[0];for(var n=[],r=1;r<=8;r++){e%2&&n.push(r),e=Math.floor(e/2)}return n},e.prototype.fromBraille=function(t){return e.unicodeToNumber(t).join("")},e.prototype.setupMap=function(){for(var t=0;t<=255;t++)this.translate.set(String.fromCodePoint(10240+t),"T")},e}(y);e.Braille2Numeric=w;var g=function(t){function e(){var e=null!==t&&t.apply(this,arguments)||this;return e.trans={0:" ",1:"f",2:"d",3:"s",4:"j",5:"k",6:"l",7:"",8:""},e}return o(e,t),e.prototype.kind=function(){return"sdf/jkl"},e.prototype.fromBraille=function(t){var e=this;return w.unicodeToNumber(t).map((function(t){return e.trans[t]})).join("")},e.prototype.setupMap=function(){for(var t=0;t<=63;t++)this.translate.set(String.fromCodePoint(10240+t),"T")},e.prototype.via=function(e){return t.prototype.via.call(this,e)+","},e}(w);e.Braille2Ascii=g},110:(t,e)=>{Object.defineProperty(e,"__esModule",{value:!0}),e.AbstractTransformer=void 0;var n=function(){function t(t,e){this.src=t,this.dst=e}return t.prototype.via=function(t){return t},t}();e.AbstractTransformer=n}},e={};function n(r){var o=e[r];if(void 0!==o)return o.exports;var i=e[r]={exports:{}};return t[r].call(i.exports,i,i.exports,n),i.exports}var r={};return(()=>{var t=r;Object.defineProperty(t,"__esModule",{value:!0}),t.Signin=t.User=t.Select=t.Convert=void 0;var e=n(362),o=n(139),i=n(277),a=n(967);t.Convert=e,t.Select=o,t.User=a,t.Signin=i})(),r})()}));
+!function(t,e){"object"==typeof exports&&"object"==typeof module?module.exports=e():"function"==typeof define&&define.amd?define([],e):"object"==typeof exports?exports.Nemeth=e():t.Nemeth=e()}(this,(()=>(()=>{"use strict";var t={647:(t,e)=>{Object.defineProperty(e,"__esModule",{value:!0}),e.NemethProjectUser=e.NemethProjectPath=e.NemethProjectDocuments=e.FeedbackStatus=e.Interaction=e.TestsCollection=e.NemethCollection=e.NemethProject=e.NemethUrl=void 0,e.NemethUrl="https://nemeth-project.firebaseio.com",e.NemethProject="nemeth-project",e.NemethCollection="nemeth",e.TestsCollection="tests",e.Interaction="InteractionStatus",e.FeedbackStatus="feedback",e.NemethProjectDocuments=e.NemethProject+"::documents",e.NemethProjectPath=e.NemethProject+"::path",e.NemethProjectUser=e.NemethProject+"::user"},416:function(t,e,n){var s=this&&this.__awaiter||function(t,e,n,s){return new(n||(n=Promise))((function(i,o){function r(t){try{c(s.next(t))}catch(t){o(t)}}function a(t){try{c(s.throw(t))}catch(t){o(t)}}function c(t){var e;t.done?i(t.value):(e=t.value,e instanceof n?e:new n((function(t){t(e)}))).then(r,a)}c((s=s.apply(t,e||[])).next())}))};Object.defineProperty(e,"__esModule",{value:!0}),e.FireTest=void 0;const i=n(647),o=n(900);e.FireTest=class{constructor(t,e,n,s,i){this.db=t,this.collection=e,this.doc=n,this.getTest=s,this.setTest=i,this.update=!0,this.countTests=0,this.preparedTests=[]}getData(){return this._data}currentTest(){return this.preparedTests[this.countTests]}prepareTests(){return s(this,void 0,void 0,(function*(){this._data=yield o.downloadData(this.db,this.collection,this.doc),this.order=this._data.order,this.tests=this._data.tests;for(const t of this.order){const e=this.tests[t];e.brf="",e.unicode="",this.preparedTests.push(e)}}))}saveTest(t){return s(this,void 0,void 0,(function*(){const e=this.currentTest();let n=null;for(const s of Object.keys(t))e[s]!==t[s]&&(n=2),e[s]=t[s];2===n&&(e[i.Interaction]=n,this.update&&o.updateData(this.db,this.collection,this.doc,e,["tests",e.name])),n||e[i.Interaction]||(e[i.Interaction]=1,this.update&&o.updateData(this.db,this.collection,this.doc,e[i.Interaction],["tests",e.name,i.Interaction]))}))}saveFeedback(t){return s(this,void 0,void 0,(function*(){const e=this.currentTest();o.updateData(this.db,this.collection,this.doc,t,["tests",e.name,i.FeedbackStatus])}))}cycleTests(t){return s(this,void 0,void 0,(function*(){yield this.saveTest(this.getTest()),this.setTest(this.nextTest(t))}))}cycleUnchangedTests(t){return s(this,void 0,void 0,(function*(){yield this.saveTest(this.getTest()),this.setTest(this.jumpTest(t,(t=>2!==t)))}))}cycleNewTests(t){return s(this,void 0,void 0,(function*(){yield this.saveTest(this.getTest()),this.setTest(this.jumpTest(t,(t=>0===t)))}))}goTest(t){return s(this,void 0,void 0,(function*(){yield this.saveTest(this.getTest());const e=this.order.indexOf(t);-1!==e&&(this.countTests=e,this.setTest(this.currentTest()))}))}nextTest(t){return this.countTests=this.countTests+(t?1:-1),this.countTests<0&&(this.countTests=this.preparedTests.length-1),this.countTests>=this.preparedTests.length&&(this.countTests=0),this.currentTest()}jumpTest(t,e){const n=this.countTests;for(;!e(this.nextTest(t)[i.Interaction])&&this.countTests!==n;);return n!==this.countTests?this.currentTest():this.nextTest(t)}}},900:function(t,e,n){var s=this&&this.__awaiter||function(t,e,n,s){return new(n||(n=Promise))((function(i,o){function r(t){try{c(s.next(t))}catch(t){o(t)}}function a(t){try{c(s.throw(t))}catch(t){o(t)}}function c(t){var e;t.done?i(t.value):(e=t.value,e instanceof n?e:new n((function(t){t(e)}))).then(r,a)}c((s=s.apply(t,e||[])).next())}))};Object.defineProperty(e,"__esModule",{value:!0}),e.restoreField=e.updateField=e.updateCollection=e.getPaths=e.setPath=e.downloadData=e.updateData=e.uploadData=e.verbose=void 0;const i=n(647);function o(t){e.verbose&&console.log(t)}function r(t,e,n,i){return s(this,void 0,void 0,(function*(){t.collection(e).doc(n).set(i).then((()=>{o(`Data for document ${n} successfully uploaded to collection ${e}`)})).catch((t=>{o(`Uploading data to document ${n} in collection ${e} failed with: ${t}`)}))}))}function a(t,e,n,i,r){return s(this,void 0,void 0,(function*(){const s={};s[r.join(".")]=i,t.collection(e).doc(n).update(s).then((()=>{o(`Data successfully updated in collection ${e} with path ${n}`)})).catch((t=>{o(`Updating data in collection ${e} with path ${n} failed with: ${t}`)}))}))}function c(t,e,n){return s(this,void 0,void 0,(function*(){return(yield t.collection(e).doc(n).get()).data()}))}function d(t,e,n,i=!0){return s(this,void 0,void 0,(function*(){const s=n.split("/")[0];let o=(yield t.collection(e).doc(s).get()).data();o||(o={}),o[n]=i,r(t,e,s,o)}))}function u(t,e,n){return s(this,void 0,void 0,(function*(){const s=n.split("/")[0];return(yield t.collection(e).doc(s).get()).data()}))}function l(t){for(const e of Object.values(t.tests))h(e)}function h(t){t[i.Interaction]=0}function f(t){for(const e of Object.values(t.tests))p(e)}function p(t){t[i.FeedbackStatus]=0}e.verbose=!1,e.uploadData=r,e.updateData=a,e.downloadData=c,e.setPath=d,e.getPaths=u,e.updateCollection=function(t,e,n,i,o=((t,e)=>{})){return s(this,void 0,void 0,(function*(){const s=[],m=yield u(t,e,i);if(!m)return s;const v=yield u(t,n,i);for(const[i,[u,T]]of Object.entries(m)){if(o(i,Object.keys(m)),s.push({name:u,information:T,path:i}),v&&v[i]&&!i.match(/^nemeth\/rules/))continue;const b=yield c(t,e,i),g=yield c(t,n,i);if(!g){l(b),f(b),yield r(t,n,i,b),yield d(t,n,i);continue}const y=g.tests;for(const e of b.order){if(y[e])continue;const s=b.tests[e];h(s),p(s),yield a(t,n,i,s,["tests",e])}yield a(t,n,i,b.order,["order"])}return s}))},e.updateField=function(t,e,n,i,o){return s(this,void 0,void 0,(function*(){const s=yield c(t,e,n);for(const r of s.order)yield a(t,e,n,o,["tests",r,i])}))},e.restoreField=function(t,e,n,i,o){return s(this,void 0,void 0,(function*(){const s=yield c(t,e,i),r=yield c(t,n,i);if(r)for(const e of r.order){const r=s.tests[e];if(!r)continue;const c=r[o];void 0!==c&&(yield a(t,n,i,c,["tests",e,o]))}}))}},986:function(t,e){var n=this&&this.__awaiter||function(t,e,n,s){return new(n||(n=Promise))((function(i,o){function r(t){try{c(s.next(t))}catch(t){o(t)}}function a(t){try{c(s.throw(t))}catch(t){o(t)}}function c(t){var e;t.done?i(t.value):(e=t.value,e instanceof n?e:new n((function(t){t(e)}))).then(r,a)}c((s=s.apply(t,e||[])).next())}))};Object.defineProperty(e,"__esModule",{value:!0}),e.harvest=e.init=e.updateAccess=e.access=e.backwardBtn=e.forwardBtn=void 0;class s{constructor(t){this.direction=t,this.span=s.createElement("btns"),this.f=s.createElement("btn f",this.direction+" element"),this.ff=s.createElement("btn ff",this.direction+" unchanged element"),this.fff=s.createElement("btn fff",this.direction+" unviewed element"),this.span.id=t;const e="next"===t?[this.f,this.ff,this.fff]:[this.fff,this.ff,this.f];for(const n of e)n.classList.add(t),n.setAttribute("tabindex","0"),this.span.appendChild(n);this.hide()}static createElement(t,e=""){const n=document.createElement("span");return n.className=t,e&&(n.title=e,n.setAttribute("aria-label",e)),n}show(){this.f.classList.remove("hidden"),this.ff.classList.remove("hidden"),this.fff.classList.remove("hidden")}hide(){this.f.classList.add("hidden"),this.ff.classList.add("hidden"),this.fff.classList.add("hidden")}}function i(t,e){t.addEventListener("click",e),t.addEventListener("keyup",(t=>{"Enter"===t.key&&e(t)}))}function o(t){i(e.backwardBtn.f,(()=>n(this,void 0,void 0,(function*(){return yield t.cycleTests(!1)})))),i(e.backwardBtn.ff,(()=>n(this,void 0,void 0,(function*(){return yield t.cycleUnchangedTests(!1)})))),i(e.backwardBtn.fff,(()=>n(this,void 0,void 0,(function*(){return yield t.cycleNewTests(!1)})))),i(e.forwardBtn.f,(()=>n(this,void 0,void 0,(function*(){return yield t.cycleTests(!0)})))),i(e.forwardBtn.ff,(()=>n(this,void 0,void 0,(function*(){return yield t.cycleUnchangedTests(!0)})))),i(e.forwardBtn.fff,(()=>n(this,void 0,void 0,(function*(){return yield t.cycleNewTests(!0)}))))}e.forwardBtn=new s("next"),e.backwardBtn=new s("previous"),e.access=new class{constructor(){this.anchor=s.createElement("btn center","access"),this.select=document.createElement("select"),this.select.id="access",this.select.classList.add("access"),this.anchor.appendChild(this.select)}init(t){this.test=t,this.select.addEventListener("change",(t=>{this.test.goTest(t.target.value)}));for(const t of this.test.order){const e=document.createElement("option");e.setAttribute("value",t),e.classList.add("access"),e.textContent=t,this.select.appendChild(e)}this.update()}update(){this.select.value=this.test.currentTest().name}},e.updateAccess=function(){e.access.update()},e.init=function(t){o(t);const n=document.createElement("div");n.classList.add("btns"),document.body.appendChild(n),n.appendChild(e.backwardBtn.span),e.backwardBtn.show(),e.access.init(t),n.appendChild(e.access.anchor);const r=s.createElement("btn center","selection");r.textContent="Selection",r.setAttribute("tabindex","0"),i(r,(()=>{t.saveTest(t.getTest()),window.location.assign("selection.html")})),n.appendChild(r),n.appendChild(e.forwardBtn.span),e.forwardBtn.show()},e.harvest=function(t,n=(()=>{})){o(t);const r=document.createElement("div");r.classList.add("btns"),document.body.appendChild(r),r.appendChild(e.backwardBtn.span),e.backwardBtn.show(),e.access.init(t),r.appendChild(e.access.anchor);const a=s.createElement("btn center","selection");a.textContent="Save",a.setAttribute("tabindex","0"),i(a,(()=>{t.saveTest(t.getTest()),n()})),r.appendChild(a),r.appendChild(e.forwardBtn.span),e.forwardBtn.show()}},362:function(t,e,n){var s=this&&this.__awaiter||function(t,e,n,s){return new(n||(n=Promise))((function(i,o){function r(t){try{c(s.next(t))}catch(t){o(t)}}function a(t){try{c(s.throw(t))}catch(t){o(t)}}function c(t){var e;t.done?i(t.value):(e=t.value,e instanceof n?e:new n((function(t){t(e)}))).then(r,a)}c((s=s.apply(t,e||[])).next())}))};Object.defineProperty(e,"__esModule",{value:!0}),e.generateMath=e.initHarvest=e.harvest=e.changeFeedback=e.changeFormat=e.generatem=e.generate=e.init=e.fireTest=void 0;const i=n(647),o=n(416),r=n(532),a=n(986),c=n(243),d=new Map([["NABT",new r.Nabt2Unicode],["BLDT",new r.Bldt2Unicode],["NABT-TABLE",new r.Nabt2UnicodeTable],["BLDT-TABLE",new r.Bldt2UnicodeTable],["SDF/JKL",new r.Ascii2Braille("","")],["1-6",new r.Numeric2Braille("","")]]),u=new Map([["NABT",new r.Unicode2Nabt],["BLDT",new r.Unicode2Bldt],["SDF/JKL",new r.Braille2Ascii("","")],["1-6",new r.Braille2Numeric("","")]]);let l="NABT";function h(){return d.get(l.toUpperCase())}const f={};e.fireTest=null;let p="";function m(t){var e;f.mathname.innerHTML=t.name,f.mathexpression.innerHTML=t.input?`<math display="block">${t.input}</math>`:t.tex?`\\[${t.tex}\\]`:`${t.name}`,f.braille.innerHTML=t.expected,f.input.value=(console.log(8),console.log(l.toUpperCase()),u.get(l.toUpperCase())).via(t.expected),f.input.focus(),function(t){if(f.refname.innerHTML="",f.references.innerHTML="",!t)return;const e=Object.keys(t).sort(((t,e)=>{const n=parseInt(t.split("-")[1],10),s=parseInt(e.split("-")[1],10);return n<s?-1:n>s?1:0})),n=e.length;if(!n)return;f.refname.innerHTML="References:";const s=[],i=n>10?T:v;for(const n of e){const e=t[n];s.push(i(n,e))}if(n<=10)return void f.references.append(...s);const o=document.createElement("select");o.setAttribute("onChange",'window.open(this.value, "_blank")'),o.classList.add("access");const r=document.createElement("option");r.setAttribute("disabled",""),r.setAttribute("selected",""),r.textContent="Select:",o.appendChild(r),o.append(...s),f.references.appendChild(o)}(t.reference),b(t[i.Interaction]),void 0!==(e=t[i.FeedbackStatus])&&(f.feedback.value=e.toString()),MathJax.typeset&&MathJax.typeset(),(0,a.updateAccess)()}function v(t,e){const n=document.createElement("a");return n.classList.add("reflink"),n.setAttribute("href",e),n.setAttribute("target","_blank"),n.textContent=t,n}function T(t,e){const n=document.createElement("option");return n.classList.add("access"),n.textContent=t,n.value=e,n}function b(t){if(void 0!==t&&f.statuscolor&&f.statusvalue)switch(t){case 0:f.statuscolor.className="green",f.statusvalue.innerHTML="New";break;case 1:f.statuscolor.className="yellow",f.statusvalue.innerHTML="Viewed";break;case 2:f.statuscolor.className="red",f.statusvalue.innerHTML="Changed";break;default:f.statuscolor.className="",f.statusvalue.innerHTML=""}}function g(){return{expected:f.braille.innerHTML}}function y(t,n){firebase?function(t,n){s(this,void 0,void 0,(function*(){const s=firebase.app().firestore();e.fireTest=new o.FireTest(s,t,n,g,m),yield e.fireTest.prepareTests(),w(),(0,a.init)(e.fireTest),e.fireTest.setTest(e.fireTest.currentTest())}))}(t,n):setTimeout(y,100)}function w(){const t=document.evaluate("//*[@id]",document.body);for(let e=t.iterateNext();e;e=t.iterateNext())f[e.id]=e}function B(){const t=f.input;p!==t.value&&("SDF/JKL"===l.toUpperCase()&&function(){if(x)return;x=!0,setTimeout((()=>{x=!1,N();const t=f.input;","!==t.value[t.value.length-1]&&(t.value=t.value+",")}),100)}(),N())}e.init=function(){const t=c.getStorage(i.NemethProjectPath),e=c.getStorage(i.NemethProjectUser);t&&e&&y(e,t)},e.generate=B;let x=!1;function N(){const t=f.input,e=t.selectionStart;f.braille.innerHTML="",f.error.innerHTML="";let[n,s,i]=function(t){const[e,n]=h().cleanInput(t);return[e,h().via(e),n]}(t.value);i&&(f.error.innerHTML="Unknown "+(i.length>1?"elements ":"element ")+i),n!==p&&b(2),t.value=n,p=n,s=s.split("\n").join("<br>"),f.braille.innerHTML=s,t.selectionEnd=e-i.length}function j(){f.mathinput.value="",f.input.value="",C(),B()}function k(t,n){return s(this,void 0,void 0,(function*(){const s=firebase.app().firestore();e.fireTest=new o.FireTest(s,t,n,g,m),yield e.fireTest.prepareTests(),w(),(0,a.harvest)(e.fireTest,j),console.log(n),console.log(t)}))}e.generatem=function(){setTimeout(B,100)},e.changeFormat=function(){e.fireTest.saveTest(e.fireTest.getTest()),l=f.format.value,e.fireTest.setTest(e.fireTest.currentTest())},e.changeFeedback=function(){const t=parseInt(f.feedback.value,10);e.fireTest.currentTest()[i.FeedbackStatus]=t,e.fireTest.saveFeedback(t)},e.harvest=function(){const t=c.getStorage(i.NemethProjectPath),e=c.getStorage(i.NemethProjectUser);t&&e&&k(e,t)},e.initHarvest=k;let L="";function C(){const t=f.mathinput;var e;L!==t.value&&(L=t.value,e=t.value,f.mathexpression.innerHTML=`\\[${e}\\]`,MathJax.typeset&&MathJax.typeset())}e.generateMath=C},243:(t,e)=>{Object.defineProperty(e,"__esModule",{value:!0}),e.setStorage=e.getStorage=e.getFirebase=void 0,e.getFirebase=function(){return firebase},e.getStorage=function(t){return localStorage.getItem(t)},e.setStorage=function(t,e){localStorage.setItem(t,e)}},139:(t,e,n)=>{Object.defineProperty(e,"__esModule",{value:!0}),e.getUid=e.addDocuments=e.addSection=void 0;const s=n(647),i=n(243);function o(t,e){const n=i.getStorage(s.NemethProjectDocuments);if(n){!function(t,e){t.sort(((t,e)=>(""+t.name).localeCompare(e.name)));for(const n of t){const t=document.createElement("tr");t.setAttribute("tabindex","0"),t.classList.add("selection"),t.innerHTML=`<td class="selection">${n.name}</td><td>${n.information}</td><td>${n.path}</td>`,t.addEventListener("click",(()=>{i.setStorage(s.NemethProjectPath,n.path),i.setStorage(s.NemethProjectUser,r()),window.location.assign("brf2nemeth.html")})),t.addEventListener("keyup",(t=>{"Enter"===t.key&&(i.setStorage(s.NemethProjectPath,n.path),i.setStorage(s.NemethProjectUser,r()),window.location.assign("brf2nemeth.html"))})),e.appendChild(t)}}(JSON.parse(n).filter((t=>t.path.match(e))),t)}}function r(){return i.getFirebase().auth().getUid()}e.addSection=function(t,e,n,s){n=`Select ${n} to work on:`;const i=document.createElement("hr");i.setAttribute("style","width:100%;border-width:0;background-color:#B1DF84;color:#B1DF84;height:6px");const r=document.createElement("div");r.classList.add("converter"),r.id=t+"table";const a=document.createElement("div");a.classList.add("section"),a.textContent=n;const c=function(t,e){const n=document.createElement("table");n.id=t;const s=document.createElement("tr");return s.classList.add("heading"),s.innerHTML="<th>Name</th><th>Information</th><th>Path</th>",n.appendChild(s),o(n,e),n}(t,e);r.append(a,c),s.append(i,r)},e.addDocuments=o,e.getUid=r},277:function(t,e,n){var s=this&&this.__awaiter||function(t,e,n,s){return new(n||(n=Promise))((function(i,o){function r(t){try{c(s.next(t))}catch(t){o(t)}}function a(t){try{c(s.throw(t))}catch(t){o(t)}}function c(t){var e;t.done?i(t.value):(e=t.value,e instanceof n?e:new n((function(t){t(e)}))).then(r,a)}c((s=s.apply(t,e||[])).next())}))};Object.defineProperty(e,"__esModule",{value:!0}),e.init=void 0;const i=n(647),o=n(900),r=n(243);function a(t){return s(this,void 0,void 0,(function*(){const e=document.createElement("span");e.textContent="Updating.",document.body.appendChild(e),yield function(t,e,n,a){return s(this,void 0,void 0,(function*(){let s=0;const c=document.createElement("div");document.body.appendChild(c);const d=yield o.updateCollection(t,e,n,a,((t,e)=>{const n=document.createElement("span");n.textContent=".",document.body.insertBefore(n,c),c.textContent=`File (${++s}/${e.length})`}));r.setStorage(i.NemethProjectDocuments,JSON.stringify(d))}))}(r.getFirebase().firestore(),i.TestsCollection,t.user.uid,i.NemethCollection),window.location.assign("selection.html")}))}const c={callbacks:{signInSuccessWithAuthResult:function(t){return a(t),!1},uiShown:function(){document.getElementById("loader").style.display="none"}},signInSuccessUrl:"brf2nemeth.html",signInOptions:[r.getFirebase().auth.EmailAuthProvider.PROVIDER_ID],tosUrl:"privacy.html",privacyPolicyUrl:"privacy.html"};e.init=function(t,e,n=c){new t.auth.AuthUI(r.getFirebase().auth()).start(e,n)}},967:(t,e)=>{function n(t){document.getElementById("user-signed-in").style.display="inline",document.getElementById("name").textContent=t.displayName,document.getElementById("email").textContent=t.email}function s(){document.getElementById("user-signed-in").style.display="none",window.location.replace(".")}function i(){window.confirm("Are you sure you want to delete your account?")&&firebase.auth().currentUser.delete().catch((t=>{"auth/requires-recent-login"===t.code&&firebase.auth().signOut().then((function(){setTimeout((function(){alert("Please sign in again to delete your account.")}),1)}))}))}function o(){const t=document.createElement("span");let e='<span id="user-signed-in" class="hidden">';e+='<span id="user-info"><span id="name"></span>',e+='(<span id="email"></span>)<div>',e+='<span><button id="sign-out">Sign Out</button></span>',e+='<span><button id="delete-account">Delete account</button></span>',e+="</div></span></span>",t.innerHTML='<span id="user-signed-in" class="hidden"><span id="user-info"><span id="name"></span>(<span id="email"></span>)<div><span><button id="sign-out">Sign Out</button></span><span><button id="delete-account">Delete account</button></span></div></span></span>',document.body.appendChild(t),document.getElementById("sign-out").addEventListener("click",(function(){firebase.auth().signOut()})),document.getElementById("delete-account").addEventListener("click",(function(){i()}))}Object.defineProperty(e,"__esModule",{value:!0}),e.init=e.addSpan=e.deleteAccount=e.handleSignedOutUser=e.handleSignedInUser=void 0,e.handleSignedInUser=n,e.handleSignedOutUser=s,e.deleteAccount=i,e.addSpan=o,e.init=function(){o(),firebase.auth().onAuthStateChanged((t=>{t?n(t):s()}))}},532:(t,e,n)=>{Object.defineProperty(e,"__esModule",{value:!0}),e.Braille2Ascii=e.Braille2Numeric=e.Ascii2Braille=e.Numeric2Braille=e.Unicode2Bldt=e.Unicode2Nabt=e.Bldt2UnicodeTable=e.Nabt2UnicodeTable=e.Bldt2Unicode=e.Nabt2Unicode=void 0;const s=n(110);class i extends s.AbstractTransformer{constructor(t,e){super(t,e),this.translate=new Map,this.setupMap()}static getFormat(t){return t.toUpperCase().match(/^BLDT/)?i.format.BLDT:i.format.NABT}via(t){let e="";for(const n of t.split("")){const t=this.translate.get(n.toLowerCase());if(!t)throw new Error("Illegal input character: "+n);e+=t}return e}cleanInput(t){const e=[],n=[];for(const s of t.toLowerCase().split(""))void 0!==this.translate.get(s)?e.push(s):n.push(s);return[e.join(""),n.join(", ")]}}i.format={NABT:" a1b'k2l@cif/msp\"e3h9o6r^djg>ntq,*5<-u8v.%[$+x!&;:4\\0z7(_?w]#y)=",BLDT:" a1b'k2l`cif/msp\"e3h9o6r~djg>ntq,*5<-u8v.%{$+x!&;:4|0z7(_?w}#y)="};class o extends i{constructor(){super("expected",""),this.dst=this.kind()}setupMap(){let t=0;for(const e of i.getFormat(this.kind()).split(""))this.translate.set(String.fromCodePoint(10240+t++),e)}}class r extends i{constructor(){super("","expected"),this.src=this.kind()}setupMap(){let t=0;for(const e of i.getFormat(this.kind()).split(""))this.translate.set(e,String.fromCodePoint(10240+t++))}}class a extends r{kind(){return"NABT"}}e.Nabt2Unicode=a;class c extends r{kind(){return"BLDT"}}e.Bldt2Unicode=c;e.Nabt2UnicodeTable=class extends a{kind(){return"NABT-TABLE"}via(t){const e=[];for(const n of t.split("\n"))e.push(super.via(n));return console.log(e.join("\n")),e.join("\n")}cleanInput(t){const e=[],n=[];for(const s of t.split("\n")){for(const t of s.toLowerCase().split(""))void 0!==this.translate.get(t)?e.push(t):n.push(t);e.push("\n")}return e.pop(),[e.join(""),n.join(", ")]}};e.Bldt2UnicodeTable=class extends c{kind(){return"BLDT-TABLE"}via(t){const e=[];for(const n of t.split("\n"))e.push(super.via(n));return console.log(e.join("\n")),e.join("\n")}cleanInput(t){const e=[],n=[];for(const s of t.split("\n")){for(const t of s.toLowerCase().split(""))void 0!==this.translate.get(t)?e.push(t):n.push(t);e.push("\n")}return e.pop(),[e.join(""),n.join(", ")]}};e.Unicode2Nabt=class extends o{kind(){return"NABT"}};e.Unicode2Bldt=class extends o{kind(){return"BLDT"}};class d extends i{via(t){return t.split(",").filter((t=>t)).map(this.toBraille.bind(this)).join("")}}class u extends i{via(t){return t.split("").map(this.fromBraille.bind(this)).join(",")}}class l extends d{kind(){return"1-6"}static numToBraille(t){const e={};let n=0;for(const s of t)!e[s]&&s>0&&s<9&&(n+=Math.pow(2,s-1)),e[s]=!0;const s=parseInt("2800",16);return String.fromCodePoint(s+n)}toBraille(t){return l.numToBraille(t.split("").map((t=>parseInt(t,10))).filter((t=>!isNaN(t))))}setupMap(){for(let t=0;t<=8;t++)this.translate.set(t.toString(),"T");this.translate.set(",","T")}}e.Numeric2Braille=l;e.Ascii2Braille=class extends l{kind(){return"sdf/jkl"}toBraille(t){const e=[];for(const n of t.split("")){const t=parseInt(this.translate.get(n),10);isNaN(t)||e.push(t)}return l.numToBraille(e)}setupMap(){this.translate=new Map([[",","T"],[" ","0"],["f","1"],["d","2"],["s","3"],["j","4"],["k","5"],["l","6"]])}};class h extends u{kind(){return"1-6"}static unicodeToNumber(t){let e=t.codePointAt(0)-10240;if(!e)return[0];const n=[];for(let t=1;t<=8;t++){e%2&&n.push(t),e=Math.floor(e/2)}return n}fromBraille(t){return h.unicodeToNumber(t).join("")}setupMap(){for(let t=0;t<=255;t++)this.translate.set(String.fromCodePoint(10240+t),"T")}}e.Braille2Numeric=h;e.Braille2Ascii=class extends h{constructor(){super(...arguments),this.trans={0:" ",1:"f",2:"d",3:"s",4:"j",5:"k",6:"l",7:"",8:""}}kind(){return"sdf/jkl"}fromBraille(t){return h.unicodeToNumber(t).map((t=>this.trans[t])).join("")}setupMap(){for(let t=0;t<=63;t++)this.translate.set(String.fromCodePoint(10240+t),"T")}via(t){return super.via(t)+","}}},110:(t,e)=>{Object.defineProperty(e,"__esModule",{value:!0}),e.AbstractTransformer=void 0;e.AbstractTransformer=class{constructor(t,e){this.src=t,this.dst=e}via(t){return t}}}},e={};function n(s){var i=e[s];if(void 0!==i)return i.exports;var o=e[s]={exports:{}};return t[s].call(o.exports,o,o.exports,n),o.exports}var s={};return(()=>{var t=s;Object.defineProperty(t,"__esModule",{value:!0}),t.Signin=t.User=t.Select=t.Convert=void 0;const e=n(362),i=n(139),o=n(277),r=n(967);t.Convert=e,t.Select=i,t.User=r,t.Signin=o})(),s})()));
 //# sourceMappingURL=nemeth-project.js.map
