@@ -19,22 +19,17 @@
  * @author volker.sorge@gmail.com (Volker Sorge)
  */
 
-import * as EngineConst from '../../speech-rule-engine/js/common/engine_const';
-import * as System from '../../speech-rule-engine/js/common/system';
-import {
-  annotators,
-  visitors
-} from '../../speech-rule-engine/js/semantic_tree/semantic_annotations';
-import { Key } from './keycodes';
+import * as EngineConst from '../../speech-rule-engine/js/common/engine_const.js';
+import * as System from '../../speech-rule-engine/js/common/system.js';
+import { Key } from './keycodes.js';
 
-import { AbstractJsonTest } from './abstract_test';
+import { AbstractJsonTest } from './abstract_test.js';
 
 export class ApiTest extends AbstractJsonTest {
   /**
    * Feature vector for setting up the engine.
    */
   public static SETUP: { [key: string]: string } = {
-    locale: 'en',
     domain: 'mathspeak',
     style: 'default',
     modality: 'speech',
@@ -61,18 +56,9 @@ export class ApiTest extends AbstractJsonTest {
    * @override
    */
   public async setUpTest() {
-    annotators.clear();
-    visitors.clear();
-    return Promise.resolve('');
-  }
-
-  /**
-   * Sets up SRE.
-   *
-   * @param feature The feature vector for the engine.
-   */
-  public async setupEngine(feature: { [key: string]: string }) {
-    await System.setupEngine(feature || ApiTest.SETUP);
+    return System.setupEngine({
+      locale: 'en'
+    });
   }
 
   /**
@@ -93,7 +79,7 @@ export class ApiTest extends AbstractJsonTest {
     json: boolean,
     move: boolean
   ) {
-    this.setupEngine(feature);
+    System.setupEngine(feature || ApiTest.SETUP);
     expr = move ? Key.get(expr) : expr || ApiTest.QUADRATIC;
     let output = (System as any)[func](expr);
     output = output
