@@ -19,11 +19,11 @@
  * @author v.sorge@mathjax.org (Volker Sorge)
  */
 
-import { SpeechTest } from './speech_test';
-import * as DomUtil from '../../speech-rule-engine/js/common/dom_util';
-import * as AuralRendering from '../../speech-rule-engine/js/audio/aural_rendering';
-import * as SpeechGeneratorUtil from '../../speech-rule-engine/js/speech_generator/speech_generator_util';
-import * as Semantic from '../../speech-rule-engine/js/semantic_tree/semantic';
+import { SpeechTest } from './speech_test.js';
+import * as DomUtil from '../../speech-rule-engine/js/common/dom_util.js';
+import * as AuralRendering from '../../speech-rule-engine/js/audio/aural_rendering.js';
+import * as SpeechGeneratorUtil from '../../speech-rule-engine/js/speech_generator/speech_generator_util.js';
+import * as Semantic from '../../speech-rule-engine/js/semantic_tree/semantic.js';
 
 export class CollapseTest extends SpeechTest {
   /**
@@ -59,7 +59,7 @@ export class CollapseTest extends SpeechTest {
   ) {
     mml =
       (pre || '') +
-      '<maction><mtext>action</mtext>' +
+      '<maction selection="2"><mtext>action</mtext>' +
       '<mrow data-semantic-id="A" ext-id="A">' +
       mml +
       '</mrow></maction>' +
