@@ -17,15 +17,19 @@
  * @author volker.sorge@gmail.com (Volker Sorge)
  */
 
-import { Grammar } from '../../speech-rule-engine/js/rule_engine/grammar';
-import * as AlphabetGenerator from '../../speech-rule-engine/js/speech_rules/alphabet_generator';
-import * as System from '../../speech-rule-engine/js/common/system';
-import { Variables } from '../../speech-rule-engine/js/common/variables';
-import * as AuralRendering from '../../speech-rule-engine/js/audio/aural_rendering';
-import { AuditoryDescription } from '../../speech-rule-engine/js/audio/auditory_description';
+import { Grammar } from '../../speech-rule-engine/js/rule_engine/grammar.js';
+import * as MathCompoundStore from '../../speech-rule-engine/js/rule_engine/math_compound_store.js';
+import * as Alphabet from '../../speech-rule-engine/js/speech_rules/alphabet.js';
+import * as System from '../../speech-rule-engine/js/common/system.js';
+import { Variables } from '../../speech-rule-engine/js/common/variables.js';
+import * as AuralRendering from '../../speech-rule-engine/js/audio/aural_rendering.js';
+import { AuditoryDescription } from '../../speech-rule-engine/js/audio/auditory_description.js';
+import { SemanticMap } from '../../speech-rule-engine/js/semantic_tree/semantic_attr.js';
+import { SpeechRuleEngine } from '../../speech-rule-engine/js/rule_engine/speech_rule_engine.js';
+import * as DomUtil from '../../speech-rule-engine/js/common/dom_util.js';
 
 import * as fs from 'fs';
-import * as tu from '../base/test_util';
+import * as tu from '../base/test_util.js';
 
 // All files that are generated.
 const InputPath = tu.TestPath.INPUT + 'common/';
@@ -40,6 +44,12 @@ const FILES = new Map([
   [SymbolType.SIUNITS, 'si_units.json'],
   [SymbolType.CHARACTERS, 'characters.json'],
   [SymbolType.FUNCTIONS, 'functions.json']
+]);
+const TYPES = new Map([
+  [SymbolType.UNITS, 'unit'],
+  [SymbolType.SIUNITS, 'unit'],
+  [SymbolType.CHARACTERS, 'character'],
+  [SymbolType.FUNCTIONS, 'function']
 ]);
 const NemethFire = true;
 
@@ -74,14 +84,28 @@ async function getCharOutput(
     locale: loc,
     style: style
   });
-  // let grammar = {translate: true};
-  const descrs = [
-    AuditoryDescription.create(
-      { text: char },
-      { adjust: true, translate: true }
-    )
-  ];
-  return AuralRendering.finalize(AuralRendering.markup(descrs));
+    let descrs = [];
+    if (modality === 'braille') {
+      if (char.match(/^\s+$/)) {
+        // TODO: This is just a temporary fix.
+        return '⠀';
+      }
+      const node = DomUtil.parseInput('<mi></mi>');
+      node.textContent = char;
+      const evaluator = SpeechRuleEngine.getInstance().getEvaluator(
+        loc,
+        modality
+      );
+      descrs = evaluator(node);
+    } else {
+      descrs = [
+        AuditoryDescription.create(
+          { text: char },
+          { adjust: true, translate: true }
+        )
+      ];
+    }
+    return AuralRendering.finalize(AuralRendering.markup(descrs));
 }
 
 /**
@@ -98,7 +122,11 @@ async function getUnitOutput(
   style: string,
   unit: string
 ) {
-  Grammar.getInstance().pushState({ annotation: 'unit' });
+  const grammar = {
+    annotation: 'unit',
+    plural: (style === 'plural')
+  };
+  Grammar.getInstance().pushState(grammar);
   const output = await getCharOutput(dom, modality, loc, style, unit);
   Grammar.getInstance().popState();
   return output;
@@ -127,7 +155,8 @@ function getOutput(
 
 const AllConstraints: { [loc: string]: string[] } = {
   ca: ['default', 'mathspeak'],
-  en: ['default', 'mathspeak', 'clearspeak'],
+  da: ['default', 'mathspeak'],
+  en: ['default', 'mathspeak', 'clearspeak', 'chromevox'],
   es: ['default', 'mathspeak'],
   fr: ['default', 'mathspeak', 'clearspeak'],
   de: ['default', 'mathspeak', 'clearspeak'],
@@ -146,29 +175,44 @@ const AllConstraints: { [loc: string]: string[] } = {
  *
  * @param locale The locale.
  * @param keys The keys for symbols to test.
- * @param unit Are the symbols units.
+ * @param kind Are the symbols units.
  * @returns The test json structure.
  */
 async function testOutput(locale: string, keys: string[], kind: SymbolType): Promise<tu.JsonFile> {
-  const constraints = AllConstraints[locale];
-  const unit = isUnitTest(kind);
+  let constraints = AllConstraints[locale];
   if (!constraints) {
     return {};
   }
-  const output: tu.JsonFile = {};
   const modality = locale === 'nemeth' ? 'braille' : 'speech';
+  const unit = modality === 'speech' && isUnitTest(kind);
+  if (unit) {
+    constraints = ['default', 'default'];
+  }
+  const output: tu.JsonFile = {};
+  let style = 'default';
   for (const dom of constraints) {
-    const json: tu.JsonFile = initialJsonFile(locale, dom, kind, 'default');
+    const json: tu.JsonFile = initialJsonFile(locale, dom, kind, style);
+    if (unit) {
+      let grammar: {[key: string]: string|boolean} = {annotation: 'unit'};
+      if (style === 'plural') {
+        grammar['plural'] = true;
+      }
+      json.grammar = grammar;
+    }
     const tests: tu.JsonTests = {};
     for (const key of keys) {
-      if (key.match(/^_comment/)) {
+      if (key.match(/^_comment/) ||
+        (json.exclude && json.exclude?.indexOf(key) !== -1)) {
         continue;
       }
-      const result = await getOutput(dom, modality, locale, 'default', key, unit);
+      const result = await getOutput(dom, modality, locale, style, key, unit);
       tests[key] = { expected: result };
     }
     json.tests = tests;
-    output[dom] = json;
+    output[unit ? style : dom] = json;
+    if (unit) {
+      style = 'plural';
+    }
   }
   return output;
 }
@@ -179,12 +223,16 @@ function initialJsonFile(loc: string, dom: string, kind: SymbolType, style?: str
   const singular =
     kind === SymbolType.CHARACTERS ? 'character' : kind.replace(/s$/, '');
   let json: tu.JsonFile = {
+    name: `${tu.TestUtil.capitalize(dom)}${tu.TestUtil.capitalize(singular)}`,
+    domain: 'default',
     locale: loc,
+    style: 'default',
+    active: `${tu.TestUtil.capitalize(dom)}Symbols${Variables.LOCALES.get(loc)}`,
     type: singular,
     factory: 'symbol',
-    name: `${tu.TestUtil.capitalize(dom)}${tu.TestUtil.capitalize(singular)}`,
-    active: `${tu.TestUtil.capitalize(dom)}${tu.TestUtil.capitalize(type)}`,
-    domain: dom
+    base: `input/common/${FILES.get(kind)}`,
+    information: `${Variables.LOCALES.get(loc)} ${tu.TestUtil.capitalize(dom)} ${tu.TestUtil.capitalize(singular)} tests.`,
+    compare: true
   };
   if (style) {
     json.style = style;
@@ -249,15 +297,15 @@ export async function testFromLocale(locale: string, kind: SymbolType): Promise<
  * @param  kind The kind of symbol for which to generate tests.
  * @param  dir Output directory.
  */
-export function testOutputFromExtras(
+export async function testOutputFromExtras(
   locale: string,
   kind: SymbolType,
   dir = '/tmp'
 ) {
-  if (kind === SymbolType.SIUNITS || locale === 'nemeth') {
+  if (isUnitTest(kind) || locale === 'nemeth' || locale === 'euro') {
     return;
   }
-  testFromExtras(locale, kind).then((output) => {
+  return testFromExtras(locale, kind).then((output) => {
     if (!Object.keys(output.tests).length) {
       return;
     }
@@ -383,7 +431,7 @@ export function testOutputFromBase(
 ) {
   testFromBase(locale, kind).then(output => {
     for (const [_dom, json] of Object.entries(output)) {
-      writeOutputToFile(dir, json, locale, json.domain, kind);
+      writeOutputToFile(dir, json, locale, isUnitTest(kind) ? json.style : json.domain, kind);
     }});
 }
 
@@ -399,7 +447,7 @@ export function testOutputFromLocale(
 ) {
   testFromLocale(locale, kind).then(output => {
     for (const [_dom, json] of Object.entries(output)) {
-      writeOutputToFile(dir, json, locale, json.domain, kind);
+      writeOutputToFile(dir, json, locale, isUnitTest(kind) ? json.style : json.domain, kind);
     }});
 }
 
@@ -407,9 +455,9 @@ export function testOutputFromLocale(
  * Compiles test output using both the locale and base file as input and writes
  * it to the destination directory.
  *
- * @param {string} locale The locale.
- * @param {SymbolType} kind The type of symbol that is considered.
- * @param {string = '/tmp'} dir The output directory.
+ * @param locale The locale.
+ * @param kind The type of symbol that is considered.
+ * @param dir The output directory.
  */
 export async function testOutputFromBoth(
   locale: string,
@@ -417,20 +465,25 @@ export async function testOutputFromBoth(
   dir = '/tmp'
 ) {
   const output = await testFromBase(locale, kind);
-  const comp = await diffBaseVsLocale(locale, kind);
-  for (const [_dom, json] of Object.entries(output)) {
-    const [base, loc] = comp[json.domain];
-    if (Object.keys(base).length && kind !== SymbolType.CHARACTERS) {
-      json.exclude = Object.keys(base);
+  const comp = await testFromLocale(locale, kind);
+  for (const [dom, json] of Object.entries(output)) {
+    if (Object.keys(comp[dom].tests).length) {
+      Object.assign(
+        json.tests, {_commentAdded_: 'Add for locale'}, comp[dom].tests);
     }
-    Object.assign(json.tests, loc);
     if (locale === 'nemeth' && kind === SymbolType.CHARACTERS && NemethFire) {
       splitNemethForFire(dir, json);
     } else {
-      writeOutputToFile(dir, json, locale, json.domain, kind);
+      writeOutputToFile(dir, json, locale, dom, kind);
     }
   }
 }
+
+
+/**
+ * Generating Nemeth test files. These are separated so they can be more
+ * conveniently handled in the Nemeth transcriber WebApp.
+ */
 
 /**
  * @param dir
@@ -462,7 +515,7 @@ function splitNemethByFile(
   const entries = locale[`nemeth/symbols/${file}.min`] as tu.JsonTest[];
   for (const entry of entries) {
     if (entry.key) {
-      keys.push(entry.key);
+      keys.push(String.fromCodePoint(parseInt(entry.key, 16)));
     }
   }
   writeNemethSymbolOutput(dir, splitOffKeys(json, keys), 'Characters', file);
@@ -473,16 +526,15 @@ function splitNemethByFile(
  * @param json
  */
 function splitNemethByAlphabet(dir: string, json: tu.JsonTests) {
-  const intervals = AlphabetGenerator.INTERVALS;
   const byFonts: { [name: string]: tu.JsonTests } = {};
-  for (const value of Object.values(AlphabetGenerator.Font)) {
+  for (const value of Object.values(Alphabet.Font)) {
     byFonts[value as string] = {};
   }
-  for (const value of Object.values(AlphabetGenerator.Embellish)) {
+  for (const value of Object.values(Alphabet.Embellish)) {
     byFonts[value as string] = {};
   }
-  for (let i = 0, int: tu.JsonTest; (int = intervals[i]); i++) {
-    const keys = AlphabetGenerator.makeInterval(int.interval, int.subst);
+  for (const int of Alphabet.INTERVALS.values()) {
+    const keys = int.unicode;
     splitOffKeys(json, keys, byFonts[int.font]);
   }
   for (const [key, values] of Object.entries(byFonts)) {
@@ -504,12 +556,10 @@ function splitOffKeys(
   keys: string[],
   result: tu.JsonTests = {}
 ) {
-  keys.forEach(function (x: string) {
-    const letter = String.fromCodePoint(parseInt(x, 16));
+  keys.forEach(function (letter: string) {
     result[letter] = json[letter];
     delete json[letter];
   });
-  console.log(Object.keys(result).length);
   return result;
 }
 
@@ -598,8 +648,10 @@ function symbolsfromLocale(json: tu.JsonTest, kind: SymbolType): tu.JsonTest[] {
 function getNamesFor(json: tu.JsonTest, kind: SymbolType): string[] {
   const symbols = symbolsfromLocale(json, kind);
   const si = kind === SymbolType.SIUNITS;
-  const prefixes =
-    json[Object.keys(json).find((j) => j.match(/^.+\/si\/prefixes\.min/))][0];
+  const match =
+    json[Object.keys(json).find((j) => j.match(/^.+\/si\/prefixes\.min/))];
+  if (!match || !match.length) return [];
+  const prefixes = match[0];
   let result: string[] = [];
   for (const obj of symbols) {
     if (si && obj.names && obj.si) {
@@ -667,7 +719,7 @@ export async function diffBaseVsLocale(
  * @param dir
  */
 export async function allTests(dir = '/tmp/symbols') {
-  for (const loc of Variables.LOCALES) {
+  for (const loc of Variables.LOCALES.keys()) {
     for (const kind of Object.values(SymbolType)) {
       await testOutputFromBoth(loc, kind, dir);
       await testOutputFromExtras(loc, kind, dir);
@@ -683,12 +735,20 @@ export function replaceTests(dir = '/tmp/symbols') {
   for (const loc of locales) {
     const files = fs.readdirSync(`${dir}/${loc}`);
     for (const file of files) {
-      console.log(file);
-      const oldJson: tu.JsonTest = tu.TestUtil.loadJson(
-        `${tu.TestPath.EXPECTED}/${loc}/symbols/${file}`
-      );
+      let oldJson: tu.JsonTest = null;
+      try {
+        oldJson = tu.TestUtil.loadJson(
+          `${tu.TestPath.EXPECTED}/${loc}/symbols/${file}`
+        );
+      } catch (err) {
+        console.warn(`Original file ${file} does not exist. Adding new one!`);
+      }
       const newJson: tu.JsonTest = tu.TestUtil.loadJson(`${dir}/${loc}/${file}`);
-      oldJson.tests = newJson.tests;
+      if (oldJson) {
+        oldJson.tests = newJson.tests;
+      } else {
+        oldJson = newJson;
+      }
       tu.TestUtil.saveJson(
         `${tu.TestPath.EXPECTED}/${loc}/symbols/${file}`,
         oldJson
@@ -696,3 +756,117 @@ export function replaceTests(dir = '/tmp/symbols') {
     }
   }
 }
+
+export async function symbolsBase() {
+  const outputBase = (kind: SymbolType, tests: tu.JsonTests) => {
+    let json = {type: TYPES.get(kind), tests: tests};
+    tu.TestUtil.saveJson(InputPath + FILES.get(kind), json);
+  }
+  await System.engineReady();
+  const chars: tu.JsonTests = {};
+  const funcs: tu.JsonTests = {};
+  const units: tu.JsonTests = {};
+  const si: tu.JsonTests = {};
+  for (const [key, value] of MathCompoundStore.subStores.entries()) {
+    if (key.match(/:unit$/)) {
+      (value.base?.si ? si : units)[key.replace(/:unit$/, '')] = {};
+      continue;
+    }
+    (value.base?.names ? funcs : chars)[key] = {};
+  }
+  outputBase(SymbolType.UNITS, units);
+  outputBase(SymbolType.SIUNITS, si);
+  outputBase(SymbolType.CHARACTERS, chars);
+  outputBase(SymbolType.FUNCTIONS, funcs);
+}
+
+/**
+ * Synchronises a semantic map test with the actual content of the map in SRE.
+ *
+ * Note, new tests still need to be filled with values using the `addFailed` method.
+ *
+ * @param expected The JSON file relative to the expected directory.
+ */
+export async function completeSemanticMapTests(expected: string) {
+  const json = tu.TestUtil.loadJson(tu.TestPath.EXPECTED + expected);
+  const jsonTests =  json.tests as tu.JsonTests;
+  const tests = Object.assign({}, jsonTests);
+  if (!tests) return;
+  const map = (SemanticMap as any)[json.map];
+  for (const key of map.keys()) {
+    if (tests[key]) {
+      delete tests[key];
+      continue;
+    }
+    jsonTests[key] = {};
+  }
+  for (const key of Object.keys(tests)) {
+    console.warn(`Removing key ${key} from mapping ${json.map}`);
+    delete jsonTests[key];
+  }
+  tu.TestUtil.saveJson(tu.TestPath.EXPECTED + expected, json);
+}
+
+/**
+ *
+ * File generators for Alphabet related tests.
+ *
+ */
+
+/**
+ * Generates base file for all automatically generated alphabets.
+ */
+export function alphabetsBase() {
+  const result: tu.JsonTests = {};
+  const fonts = Object.values(Alphabet.Font) as string[];
+  const embel = Object.values(Alphabet.Embellish) as string[];
+  for (const font of fonts.concat(embel)) {
+    for (const base of Object.values(Alphabet.Base)) {
+      const interval = Alphabet.INTERVALS.get(Alphabet.alphabetName(base, font));
+      if (!interval) continue;
+      const tag = base === Alphabet.Base.DIGIT ? 'mn' : 'mi';
+      interval.unicode.forEach(x =>
+        result[x] = {input: `<${tag}>${x}</${tag}>`});
+    }
+  }
+  tu.TestUtil.saveJson(InputPath + 'alphabets.json', {tests: result});
+}
+
+/**
+ * Generates the skeleton expected file for a locale. This still needs to be
+ * filled with values using the `addMissing` method.
+ *
+ * @param locale The iso string for the locale.
+ */
+export function alphabetsExpected(locale: string) {
+  const constraints = AllConstraints[locale];
+  if (!constraints) {
+    return;
+  }
+  const modality = locale === 'nemeth' ? 'braille' : 'speech';
+  const loc = Variables.LOCALES.get(locale);
+  for (const dom of constraints) {
+    if (dom === 'default' && modality === 'speech') continue;
+    let json: tu.JsonFile = {
+      locale: locale,
+      factory: 'speech',
+      name: `${tu.TestUtil.capitalize(loc)}${tu.TestUtil.capitalize(dom)}Alphabets`,
+      active: `Alphabets${tu.TestUtil.capitalize(loc)}`,
+      information: `${tu.TestUtil.capitalize(loc)} ${tu.TestUtil.capitalize(dom)} Alphabet speech tests.`,
+      domain: dom,
+      modality: modality,
+      base: 'input/common/alphabets.json',
+      tests: {}
+    }
+    tu.TestUtil.saveJson(`${tu.TestPath.EXPECTED}${locale}/${modality === 'speech' ? dom : 'rules'}/alphabets.json`, json);
+  }
+};
+
+/**
+ * Generates skeleton expected files for all available locales.
+ */
+export function alphabetsAllExpected() {
+  for (const loc of Variables.LOCALES.keys()) {
+    alphabetsExpected(loc);
+  }
+};
